% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{ab4B}
\alias{ab4B}
\title{ab4B}
\format{A data frame with 251 observations on the following 5 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector} \item{list("CAS")}{a factor with
odorant CAS numbers} \item{list("Bruyne.2001.RR")}{a numeric vector;
electrophysiological recording data of response range from de Bruyne et al.
2001 \link[=reference]{(de Bruyne et al., 2001)}} }}
\description{
Response profile of ab4B
}
\details{
\describe{ \item{list("coexpression")}{ Or33a coexpress with Or56a in ab4B
neuron \link[=reference]{(Fishilevich et al., 2005)} and coconvergence to
glomerulus DA2 \link[=reference]{(Fishilevich et al., 2005)}.  }\item{:}{
Or33a coexpress with Or56a in ab4B neuron \link[=reference]{(Fishilevich et
al., 2005)} and coconvergence to glomerulus DA2
\link[=reference]{(Fishilevich et al., 2005)}.  } }
}
\examples{
data(ab4B)
}
\references{
\itemize{ \item \link[=reference]{(de Bruyne et al., 2001)} : de
Bruyne M., Foster K., Carlson J. R., 2001, \emph{Odor coding in the
Drosophila antenna, Neuron}, 30:537-552 }
}
\keyword{datasets}

