% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{ab5B}
\alias{ab5B}
\title{ab5B}
\format{A data frame with 251 observations on the following 7 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector} \item{list("CAS")}{a factor with
odorant CAS numbers} \item{list("Hallem.2004.WT")}{a numeric vector;
electrophysiological recording in wild type from Hallem et al. 2004
\link[=reference]{(Hallem et al., 2004)} } \item{list("Bruyne.2001.WT")}{a
numeric vector; electrophysiological recording in wild type from de Bruyne
et al. 2001 \link[=reference]{(de Bruyne et al., 2001)} }
\item{list("Schmuker.2007.TR")}{a numeric vector; electrophysiological
recording data from Schmuker et al. 2007 \link[=reference]{(Schmuker et al.,
2007)} } }}
\description{
Response profile of ab5B
}
\details{
\describe{ \item{list("coexpression")}{ Or47a coexpress with Or33b in ab5B
neuron \link[=reference]{(Fishilevich et al., 2005)} and coconvergence to
glomerulus DM3 \link[=reference]{(Couto et al., 2005)},
\link[=reference]{(Fishilevich et al., 2005)}.  }\item{:}{ Or47a coexpress
with Or33b in ab5B neuron \link[=reference]{(Fishilevich et al., 2005)} and
coconvergence to glomerulus DM3 \link[=reference]{(Couto et al., 2005)},
\link[=reference]{(Fishilevich et al., 2005)}.  } }
}
\examples{
data(ab5B)
}
\references{
\itemize{ \item \link[=reference]{(Hallem et al., 2004)} :
Hallem E.A., Ho M. G. and Carlson J.R., 2004, \emph{The Molecular Basis of
Odor Coding in the Drosophila Antenna}, Cell, 117:965-979 \item
\link[=reference]{(de Bruyne et al., 2001)} : de Bruyne M., Foster K.,
Carlson J. R., 2001, \emph{Odor coding in the Drosophila antenna, Neuron},
30:537-552 \item \link[=reference]{(Schmuker et al., 2007)} : Schmuker M.,
de Bruyne M., Haehnel M., Schneider G., 2007, \emph{Predicting olfactory
receptor neuron responses from odorant structure},Chem Cent J., 1:11 }
}
\keyword{datasets}

