% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{ac1}
\alias{ac1}
\title{ac1}
\format{A data frame with 693 observations on 6 variables.}
\description{
DoOR response data
}
\details{
pooled measurements of the ac1 sensillum
}
\keyword{datasets}

