% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{ac1b}
\alias{ac1b}
\title{ac1b}
\format{A data frame with 251 observations on the following 5 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector} \item{list("CAS")}{a factor with
odorant CAS numbers} \item{list("Yao.2005.WT")}{a numeric vector;
electrophysiological recording in wild type from Yao et al. 2005
\link[=reference]{(Yao et al., 2005)} } }}
\description{
Response profile of ac1b
}
\details{
ac1b \describe{ \item{list("Sensilla")}{ antennal coeloconic sensilla
(antenna) \link[=reference]{(Yao et al., 2005)} }\item{:}{ antennal
coeloconic sensilla (antenna) \link[=reference]{(Yao et al., 2005)} }
\item{list("Neuron")}{ ac1b \link[=reference]{(Yao et al., 2005)} }\item{:}{
ac1b \link[=reference]{(Yao et al., 2005)} } \item{list("Comment")}{
\enumerate{ \item ac1b is excited by none of 45 tested odors
\link[=reference]{(Yao et al., 2005)}.  \item It seems likely that ac1b
sensitive to humidity \link[=reference]{(Yao et al., 2005)}.  } }\item{:}{
\enumerate{ \item ac1b is excited by none of 45 tested odors
\link[=reference]{(Yao et al., 2005)}.  \item It seems likely that ac1b
sensitive to humidity \link[=reference]{(Yao et al., 2005)}.  } } }
}
\examples{
data(ac1b)
}
\references{
\itemize{ \item \link[=reference]{(Yao et al., 2005)} : Yao C.
A., Ignell, R. and Carlson J. R., \emph{Chemosensory Coding by Neurons in
the Coeloconic Sensilla of the Drosophila Antenna}, 2005, 25(37):8359-8367 }
}
\keyword{datasets}

