% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{ac1BC}
\alias{ac1BC}
\title{ac1BC}
\format{A data frame with 693 observations on 6 variables.}
\description{
DoOR response data
}
\details{
pooled measurements of ac1B and ac1C OSNs
}
\keyword{datasets}

