% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{ac2}
\alias{ac2}
\title{ac2}
\format{A data frame with 693 observations on 6 variables.}
\description{
DoOR response data
}
\details{
pooled measurements of the ac2 sensilum
}
\keyword{datasets}

