% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{ac2BC}
\alias{ac2BC}
\title{ac2BC}
\format{A data frame with 693 observations on 6 variables.}
\description{
DoOR response data
}
\details{
pooled measurements of ac2B and ac2C OSNs
}
\keyword{datasets}

