% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{ac3_noOr35a}
\alias{ac3_noOr35a}
\title{ac3_noOr35a}
\format{A data frame with 693 observations on 6 variables.}
\description{
DoOR response data
}
\details{
mpooled measurements of the ac3 sensillum with Or35a silenced
}
\keyword{datasets}

