% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{AL.map}
\alias{AL.map}
\title{AL.map}
\format{The format is: List of 5 $ glomeruli :'data.frame': 19834 obs. of 3
variables: ..$ glomerulus: Factor w/ 51 levels "D","DA1","DA2",..: 4 4 4 4 4
4 4 4 4 4 ...  ..$ x : num [1:19834] 33.1 33.1 33.1 33.1 33.1 ...  ..$ y :
num [1:19834] 24.2 24.2 24.3 24.4 24.4 ...  $ unmapped_not.olf:'data.frame':
1274 obs. of 3 variables: ..$ glomerulus: Factor w/ 3 levels
"VA7m","VP2","VP3": 3 3 3 3 3 3 3 3 3 3 ...  ..$ x : num [1:1274] 280 279
279 279 279 ...  ..$ y : num [1:1274] 33.8 33.8 33.9 33.9 34 ...  $
background :'data.frame': 9428 obs. of 3 variables: ..$ group: Factor w/ 4
levels "slice1","slice2",..: 4 4 4 4 4 4 4 4 4 4 ...  ..$ x : num [1:9428]
39 38.9 38.8 38.7 38.6 ...  ..$ y : num [1:9428] 62.2 62.2 62.2 62.1 62.1
...  $ bg.cutout :'data.frame': 2361 obs. of 3 variables: ..$ group: Factor
w/ 5 levels "c1","c2","c3",..: 5 5 5 5 5 5 5 5 5 5 ...  ..$ x : num [1:2361]
99.1 99.3 99.5 99.6 99.8 ...  ..$ y : num [1:2361] 15.6 15.6 15.6 15.6 15.7
...  $ labels :'data.frame': 54 obs. of 3 variables: ..$ glomerulus: Factor
w/ 54 levels "D","DA1","DA2",..: 34 27 12 5 32 3 52 51 54 53 ...  ..$ x :
num [1:54] 157 105 110 31 25 ...  ..$ y : num [1:54] 51.9 55 15 35 41 ...}
\description{
The AL map after Grabe et al.
}
\keyword{datasets}

