\name{AL256}
\Rdversion{1.1}
\alias{AL256}
\docType{data}
\title{
AL256
}
\description{
  A 422 by 1495 Figure matrix of Antennal Lobe 
}
\usage{data(AL256)}
\format{
  A data frame with 422 observations on the following 1495 variables.
}
\references{
	Vosshall L. B. and Stocker R. F., 2007, \emph{Molecular Architecture of Smell and Taste in Drosophila}, 
	Annu. Rev. Neurosci., 30:505-533
}
\examples{
data(AL256)
}
\keyword{datasets}
