% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{door_dataset_info}
\alias{door_dataset_info}
\title{door_dataset_info}
\format{A data frame with 42 observations on 15 variables.}
\description{
information on the studies that contributed to DoOR
}
\keyword{datasets}

