% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{package}
\name{DoOR.data-package}
\alias{DoOR.data}
\alias{DoOR.data-package}
\title{DoOR Data}
\description{
Data package for DoOR.
}
\details{
\tabular{ll}{ Package: \tab DoOR.data\cr Type: \tab Package\cr Version: \tab
0.1-2\cr Date: \tab 2012-04-16\cr License: \tab GPL-2\cr LazyLoad: \tab
yes\cr }

\bold{Type \code{help(package = DoOR.data)} to see a complete list of
datasets and functions.  Below is a list of DoOR.data.}

\bold{Response Data:} \tabular{ll}{ \code{\link{ab2B}}: \tab odorant
receptor neuron ab2B (express both Or33b and Or85a); housing in antennal
basiconic sensillum \cr \code{\link{ab3B}}: \tab odorant receptor neuron
ab3B (express both Or85b and Or98b); housing in antennal basiconic sensillum
\cr \code{\link{ab4B}}: \tab odorant receptor neuron ab4B (express both
Or33a and Or56a); housing in antennal basiconic sensillum \cr
\code{\link{ab5B}}: \tab odorant receptor neuron ab5B (express both Or33b
and Or47a); housing in antennal basiconic sensillum \cr \code{\link{ac1a}}:
\tab odorant receptor ac1a; housing in antennal coeloconic sensillum \cr
\code{\link{ac1b}}: \tab odorant receptor ac1b; housing in antennal
coeloconic sensillum \cr \code{\link{ac2a}}: \tab odorant receptor ac2a;
housing in antennal coeloconic sensillum \cr \code{\link{ac2b}}: \tab
odorant receptor ac2b; housing in antennal coeloconic sensillum \cr
\code{\link{ac3a}}: \tab odorant receptor ac3a; housing in antennal
coeloconic sensillum \cr \code{\link{ac4}}: \tab odorant receptor ac4;
housing in antennal coeloconic sensillum \cr \code{\link{Ir31a}}: \tab
ionotropic glutamate receptor Ir31a; housing in antennal coeloconic
sensillum \cr \code{\link{Ir75a}}: \tab ionotropic glutamate receptor Ir75a;
housing in antennal coeloconic sensillum \cr \code{\link{Ir75d}}: \tab
ionotropic glutamate receptor Ir75d; housing in antennal coeloconic
sensillum \cr \code{\link{Ir76a}}: \tab ionotropic glutamate receptor Ir76a;
housing in antennal coeloconic sensillum \cr \code{\link{Ir76b}}: \tab
ionotropic glutamate receptor Ir76b; housing in antennal coeloconic
sensillum \cr \code{\link{Ir84a}}: \tab ionotropic glutamate receptor Ir84a;
housing in antennal coeloconic sensillum \cr \code{\link{Ir92a}}: \tab
ionotropic glutamate receptor Ir92a; housing in antennal coeloconic
sensillum \cr \code{\link{Gr21a}}: \tab gustatory receptor neuron Gr21a;
housing in antennal coeloconic sensillum \cr \code{\link{Or10a}}: \tab
odorant receptor Or10a; housing in antennal basiconic sensillum \cr
\code{\link{Or13a}}: \tab odorant receptor Or13a; housing in antennal
basiconic sensillum \cr \code{\link{Or19a}}: \tab odorant receptor Or19a;
housing in antennal trichoid sensillum \cr \code{\link{Or1a}}: \tab odorant
receptor Or1a; larval specific odorant receptor \cr \code{\link{Or22a}}:
\tab odorant receptor Or22a; housing in antennal basiconic sensillum \cr
\code{\link{Or22b}}: \tab odorant receptor Or22b; housing in antennal
basiconic sensillum \cr \code{\link{Or22c}}: \tab odorant receptor Or22c;
larval specific odorant receptor \cr \code{\link{Or23a}}: \tab odorant
receptor Or23a; housing in antennal trichoid sensillum \cr
\code{\link{Or24a}}: \tab odorant receptor Or24a; larval specific odorant
receptor \cr \code{\link{Or2a}}: \tab odorant receptor Or2a; housing in
antennal trichoid sensillum \cr \code{\link{Or30a}}: \tab odorant receptor
Or30a; larval specific odorant receptor \cr \code{\link{Or33a}}: \tab
odorant receptor Or33a; housing in antennal basiconic sensillum \cr
\code{\link{Or33b}}: \tab odorant receptor Or33b; housing in antennal
basiconic sensillum \cr \code{\link{Or33c}}: \tab odorant receptor Or33c;
housing in maxillary basiconic sensillum \cr \code{\link{Or35a}}: \tab
odorant receptor Or35a; housing in antennal coeloconic sensillum \cr
\code{\link{Or42a}}: \tab odorant receptor Or42a; housing in maxillary
basiconic sensillum \cr \code{\link{Or42b}}: \tab odorant receptor Or42b;
housing in antennal basiconic sensillum \cr \code{\link{Or43a}}: \tab
odorant receptor Or43a; housing in antennal trichoid sensillum \cr
\code{\link{Or43b}}: \tab odorant receptor Or43b; housing in antennal
basiconic sensillum \cr \code{\link{Or45a}}: \tab odorant receptor Or45a;
larval specific odorant receptor \cr \code{\link{Or45b}}: \tab odorant
receptor Or45b; larval specific odorant receptor \cr \code{\link{Or46a}}:
\tab odorant receptor Or46a; housing in maxillary basiconic sensillum \cr
\code{\link{Or47a}}: \tab odorant receptor Or47a; housing in antennal
basiconic sensillum \cr \code{\link{Or47b}}: \tab odorant receptor Or47b;
housing in antennal trichoid sensillum \cr \code{\link{Or49a}}: \tab odorant
receptor Or49a; housing in antennal basiconic sensillum \cr
\code{\link{Or49b}}: \tab odorant receptor Or49b; housing in antennal
basiconic sensillum \cr \code{\link{Or56a}}: \tab odorant receptor Or56a;
housing in antennal basiconic sensillum \cr \code{\link{Or59a}}: \tab
odorant receptor Or59a; larval specific odorant receptor \cr
\code{\link{Or59b}}: \tab odorant receptor Or59b; housing in antennal
basiconic sensillum \cr \code{\link{Or59c}}: \tab odorant receptor Or59c;
housing in maxillary basiconic sensillum \cr \code{\link{Or63a}}: \tab
odorant receptor Or63a; larval specific odorant receptor \cr
\code{\link{Or65a}}: \tab odorant receptor Or65a; housing in antennal
trichoid sensillum \cr \code{\link{Or67a}}: \tab odorant receptor Or67a;
housing in antennal basiconic sensillum \cr \code{\link{Or67b}}: \tab
odorant receptor Or67b; housing in antennal basiconic sensillum \cr
\code{\link{Or67c}}: \tab odorant receptor Or67c; housing in antennal
basiconic sensillum \cr \code{\link{Or67d}}: \tab odorant receptor Or67d;
housing in antennal trichoid sensillum \cr \code{\link{Or69a}}: \tab odorant
receptor Or69a; housing in antennal basiconic sensillum \cr
\code{\link{Or71a}}: \tab odorant receptor Or71a; housing in maxillary
basiconic sensillum \cr \code{\link{Or74a}}: \tab odorant receptor Or74a;
larval specific odorant receptor \cr \code{\link{Or7a}}: \tab odorant
receptor Or7a; housing in antennal basiconic sensillum \cr
\code{\link{Or82a}}: \tab odorant receptor Or82a; housing in antennal
basiconic sensillum \cr \code{\link{Or83a}}: \tab odorant receptor Or83a;
larval specific odorant receptor \cr \code{\link{Or83c}}: \tab odorant
receptor Or83c; housing in antennal trichoid sensillum \cr
\code{\link{Or85a}}: \tab odorant receptor Or85a; housing in antennal
basiconic sensillum \cr \code{\link{Or85b}}: \tab odorant receptor Or85b;
housing in antennal basiconic sensillum \cr \code{\link{Or85c}}: \tab
odorant receptor Or85c; larval specific odorant receptor \cr
\code{\link{Or85d}}: \tab odorant receptor Or85d; housing in maxillary
basiconic sensillum \cr \code{\link{Or85e}}: \tab odorant receptor Or85e;
housing in maxillary basiconic sensillum \cr \code{\link{Or85f}}: \tab
odorant receptor Or85f; ousing in antennal basiconic sensillum \cr
\code{\link{Or88a}}: \tab odorant receptor Or88a; housing in antennal
trichoid sensillum \cr \code{\link{Or92a}}: \tab odorant receptor Or92a;
ousing in antennal basiconic sensillum \cr \code{\link{Or94a}}: \tab odorant
receptor Or94a; larval specific odorant receptor \cr \code{\link{Or94b}}:
\tab odorant receptor Or94b; larval specific odorant receptor \cr
\code{\link{Or98a}}: \tab odorant receptor Or98a; ousing in antennal
basiconic sensillum \cr \code{\link{Or9a}}: \tab odorant receptor Or9a;
ousing in antennal basiconic sensillum \cr \code{\link{pb2A}}: \tab odorant
receptor neuron pb2A (express both Or33c and Or85e); housing in the
sensillum of maxillary palp \cr }

\bold{Supported Data:} \tabular{ll}{ \code{\link{AL256}}: \tab a 422 by 1495
Figure matrix of Antennal Lobe. \cr \code{\link{dataset.info}}: \tab dataset
information \cr \code{\link{data.format}}: \tab the format of response data
\cr \code{\link{glo.dist}}: \tab a matrix of distance coefficients between
glomeruli \cr \code{\link{odor}}: \tab database of Odors \cr
\code{\link{OGN}}: \tab mapping of Receptors \cr \code{\link{ORs}}: \tab
receptor names and expressions \cr \code{\link{reference}}: \tab reference
for response data \cr \code{\link{response.matrix}}: \tab an global
unnormalized responses matrix \cr \code{\link{response.range}}: \tab
response ranges for each study \cr
\code{\link{unglobalNorm_response.matrix}}: \tab an unnormalized responses
matrix \cr \code{\link{weight.globNorm}}: \tab weight matrix for global
normalizazion \cr }
}
\author{
C. Giovanni Galizia \cr Daniel Muench \cr Martin Strauch \cr Anja
Nissler \cr Shouwen Ma \cr

Maintainer: Daniel Muench <daniel.muench@uni-konstanz.de> Shouwen Ma
<shouwen.ma@gmail.com>
}
\references{
\url{http://neuro.uni-konstanz.de/DoOR/default.html} and See
\code{\link{reference}}
}
\seealso{
\code{\link[<pkg>:<pkg>-package]{<DoOR.function>}}
}
\keyword{package}

