% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{door_dataset_info}
\alias{door_dataset_info}
\title{door_dataset_info}
\format{'data.frame':	42 obs. of  15 variables:
$ dataset                      : Factor w/ 42 levels "Bruyne.1999.WT",..: 1 2 3 4 5 6 7 9 10 8 ...
$ study                        : Factor w/ 29 levels "de Bruyne et.al. 1999",..: 1 2 2 3 4 4 5 6 6 7 ...
$ other.dataset.in.this.study  : Factor w/ 22 levels "","Bruyne.2001.RR",..: 1 3 2 1 5 4 1 1 1 1 ...
$ other.dataset.in.this.study.2: Factor w/ 5 levels "","Pelz.2005.Antnmr",..: 1 1 1 1 1 1 1 1 1 1 ...
$ SFR.reported                 : Factor w/ 3 levels "no","no SFR",..: 3 1 3 1 1 1 1 1 1 1 ...
$ SFR.substracted              : Factor w/ 4 levels "no","no SFR",..: 4 1 4 4 1 1 4 4 4 4 ...
$ technique                    : Factor w/ 2 levels "calcium imaging",..: 2 2 2 2 2 2 2 2 2 2 ...
$ data.type                    : Factor w/ 4 levels "EC50","mean deltaF/F",..: 4 4 4 4 4 4 4 4 4 4 ...
$ control                      : Factor w/ 7 levels "","air","identical filter papers with 10 ml of the solvent.",..: 5 5 5 5 5 1 1 1 1 1 ...
$ solvents                     : Factor w/ 10 levels "","Acetoin and 1-propanethiol were diluted in water, all others odors in paraffin oil",..: 8 8 8 8 8 1 7 1 1 1 ...
$ solvents.subtracted          : Factor w/ 3 levels "","no","yes": 2 1 2 2 1 1 3 3 3 3 ...
$ concentration                : Factor w/ 5 levels "","10^-2","10^-2 -  (vol/vol)",..: 2 2 2 3 3 3 3 3 3 1 ...
$ comment                      : Factor w/ 6 levels "","Data from de Bruyne et.al.,2001.",..: 6 6 6 1 1 1 4 1 1 1 ...
$ link                         : Factor w/ 28 levels "","http://chemse.oxfordjournals.org/content/35/7/551",..: 22 13 13 6 14 14 9 28 28 27 ...
$ DOI                          : Factor w/ 26 levels "","10.1002/minf.201300037",..: 1 14 14 4 15 15 9 10 10 18 ...}
\description{
Detailed information on the sources of the data sets in used in DoOR.
}
\keyword{DoOR}
\keyword{dataset}
\keyword{door_dataset_info}
\keyword{responding_unit}

