% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{door_excluded_data}
\alias{door_excluded_data}
\title{door_excluded_data}
\format{
'data.frame':	78 obs. of  2 variables: $ OR      : Factor w/ 78 levels "ab2B","ab4B",..: 5
6 9 10 12 13 15 29 35 63 ... $ excluded: Factor w/ 10 levels "","Bruyne.2001.RR,
Marshall.2010.WT",..: 1 NA 1 NA 1 1 1 NA 8 8 ...
}
\description{
These data sets have been excluded in the current response matrices. Reasons might for example be
low overlap with the remaining sets or bad fits.
}
\keyword{DoOR}
\keyword{dataset}
\keyword{door_excluded_data}
\keyword{responding_unit}
