% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{door_glo_dist}
\alias{door_glo_dist}
\title{door_glo_dist}
\format{
'data.frame':	49 obs. of  49 variables:
$ D                : num  0 19.33 11.1 9.98 12.86 ...
$ DA1              : num  19.3 0 24.4 13.7 13.5 ...
$ DA2              : num  11.1 24.4 0 15.3 12 ...
$ DA3              : num  9.98 13.7 15.28 0 10.01 ...
$ DA4l             : num  12.9 13.5 12 10 0 ...
$ DA4m             : num  8.29 18.18 6.94 8.59 6.86 ...
$ DC1              : num  10.9 20.3 12.8 17.5 13.4 ...
$ DC2              : num  22.3 18.2 20.2 22.5 13.4 ...
$ DC3              : num  20.7 30.7 21.5 29 24.9 ...
$ DL1              : num  17.5 12.7 26.2 17.9 19.6 ...
$ DL2              : num  29.1 18 34.2 29 25.8 ...
$ DL3              : num  13.66 12.25 22.1 7.71 15.9 ...
$ DL4              : num  9.64 11.7 18.89 9.13 13.47 ...
$ DL5              : num  11.1 21 20.4 18.1 20 ...
$ DM1              : num  25.2 39.8 27 35.1 33.7 ...
$ DM2              : num  25.2 40.9 21.4 34.2 30.9 ...
$ DM3              : num  14.3 32 16.7 24.2 24.5 ...
$ DM4              : num  32.7 46.1 30.9 42 38.2 ...
$ DM5              : num  25.4 38.8 17.7 32.4 27.1 ...
$ DM6              : num  14.9 33.1 11.1 22.9 22.3 ...
$ DP1l             : num  30.3 23.9 34.8 32.6 28.7 ...
$ DP1m             : num  23.1 28.5 27.9 30.1 27.7 ...
$ extra.glomerulus1: num  16.8 28 11.8 23.6 17.7 ...
$ extra.glomerulus2: num  61.5 59.2 58.7 65.4 56.9 ...
$ V                : num  64.2 59.7 61.6 67.1 58.4 ...
$ VA1d             : num  25.1 15.3 23.2 21.4 13 ...
$ VA1lm            : num  35.8 24.7 35 34.4 26.4 ...
$ VA1v             : num  31.2 16.6 32.6 28.2 22.1 ...
$ VA2              : num  38.8 44.8 32 44.1 35.6 ...
$ VA3              : num  42 39.8 36.2 43.7 33.9 ...
$ VA4              : num  49.4 46.6 44.8 51.8 42.4 ...
$ VA5              : num  42.9 34.7 39.4 42.3 33.1 ...
$ VA6              : num  19 23.3 13.6 22.1 13.2 ...
$ VA7l             : num  33.6 29.3 28.9 34 24.2 ...
$ VA7m             : num  30.5 30.5 24.5 32.4 22.7 ...
$ VC1              : num  42.2 41.1 39.5 46.1 37.7 ...
$ VC2              : num  38.8 36.6 35.1 41.6 32.5 ...
$ VC3              : num  49.6 52.8 47.2 55.8 48.4 ...
$ VC4              : num  44.6 48 41.5 50.6 42.8 ...
$ VL1              : num  53.3 45.8 51.2 54.4 45.9 ...
$ VL2a             : num  42 29.7 43.6 41.3 35 ...
$ VL2p             : num  51.1 39.7 52.6 51.1 44.8 ...
$ VM1              : num  56.5 60 53.5 62.8 55.2 ...
$ VM2              : num  36.5 46.9 30.5 43.8 37.2 ...
$ VM3              : num  45.2 51.8 40.1 51.8 44.1 ...
$ VM4              : num  56.4 57.4 51.9 61.1 52.2 ...
$ VM5d             : num  26 35 20.1 32.4 25.2 ...
$ VM6              : num  62.7 62.9 59.4 67.6 59.3 ...
$ VM7              : num  35 43.8 33.1 43 37.4 ...
}
\description{
Spatial distances between glomeruli as reported in Laissue PP, Reiter C, Hiesinger PR, Halter S,
Fischbach KF, Stocker RF. 1999. Three-dimensional reconstruction of the antennal lobe in
Drosophila melanogaster. J Comp Neurol. 405:543–552.
}
\keyword{DoOR}
\keyword{dataset}
\keyword{door_glo_dist}
\keyword{responding_unit}
