% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{door_global_normalization_weights}
\alias{door_global_normalization_weights}
\title{door_global_normalization_weights}
\format{
'data.frame':	78 obs. of  42 variables:
$ Bruyne.2001.RR                 : int  NA NA NA NA NA NA NA NA NA 1 ...
$ Bruyne.2001.WT                 : int  NA NA NA NA NA NA NA NA NA 1 ...
$ Bruyne.2010.WT                 : int  NA NA NA NA NA NA NA NA NA 1 ...
$ Dobritsa.2003.EN               : int  NA NA NA NA NA NA NA NA NA NA ...
$ Dobritsa.2003.WT               : int  NA NA NA NA NA NA NA NA NA NA ...
$ Dweck.2013.WT                  : int  NA NA NA NA NA NA NA NA NA NA ...
$ Dweck.2015b.WT                 : int  NA NA NA NA NA NA NA NA NA NA ...
$ Dweck.2015.EN                  : int  NA NA NA NA NA NA NA NA NA NA ...
$ Dweck.2015.WT                  : int  NA NA NA NA NA NA NA NA NA NA ...
$ Gabler.2013.AL                 : int  NA NA NA NA NA NA NA NA NA 1 ...
$ Galizia.2009.nmr               : int  NA NA NA NA NA NA NA NA NA NA ...
$ Goldman.2005.EN                : int  NA NA NA NA NA NA NA NA NA NA ...
$ Goldman.2005.WT                : int  NA NA NA NA NA NA NA NA NA NA ...
$ Hallem.2006.EN                 : int  NA NA NA NA NA NA NA NA 1 1 ...
$ Kreher.2005.EN                 : int  NA NA NA NA NA NA NA NA NA NA ...
$ Kreher.2008.EN                 : int  NA NA NA NA NA NA NA 1 1 1 ...
$ Kwon.2007.EN                   : int  NA NA NA NA NA NA NA NA NA NA ...
$ Kwon.2007.WT                   : int  NA NA NA NA NA NA NA NA NA NA ...
$ Marshall.2010.WT               : int  1 NA 1 NA 1 1 1 NA NA 1 ...
$ Montague.2011.EN               : int  NA NA NA NA NA NA NA NA 1 1 ...
$ Nissler.2007.EC50              : int  NA NA NA NA NA NA NA NA NA NA ...
$ Nissler.2007.nmr               : int  NA NA NA NA NA NA NA NA NA NA ...
$ Pelz.2005.ALnmr                : int  NA NA NA NA NA NA NA NA NA NA ...
$ Pelz.2005.Antnmr               : int  NA NA NA NA NA NA NA NA NA NA ...
$ Pelz.2005.Or47bnmr             : int  NA NA NA NA NA NA NA NA NA NA ...
$ Pelz.2006.ALEC50               : int  NA NA NA NA NA NA NA NA NA NA ...
$ Pelz.2006.AntEC50              : int  NA NA NA NA NA NA NA NA NA NA ...
$ Ronderos.2014.WT               : int  NA NA NA NA NA NA NA NA NA NA ...
$ Schmuker.2007.TR               : int  NA NA NA NA NA NA NA NA NA NA ...
$ Silbering.2011.AL              : int  NA NA NA NA NA 1 NA NA NA NA ...
$ Silbering.2011.AL_8a           : int  NA NA NA NA 1 NA NA NA NA NA ...
$ Silbering.2011.WT              : int  NA NA NA NA NA NA 1 NA NA NA ...
$ Stensmyr.2003.WT               : int  NA NA NA NA NA NA NA NA NA NA ...
$ Stensmyr.2012.WT               : int  NA NA NA NA NA NA NA NA NA 1 ...
$ Turner.2009.SC                 : int  NA NA NA NA NA NA NA NA NA NA ...
$ van.der.Goes.van.Naters.2007.EN: int  NA NA NA NA NA NA NA NA NA NA ...
$ Yao.2005.WT                    : int  1 1 1 1 1 1 1 NA NA NA ...
$ Hallem.2004.EN                 : int  NA NA NA NA NA NA NA NA 1 1 ...
$ Hallem.2004.WT                 : int  NA NA NA NA NA 1 NA NA NA 1 ...
$ Muench.2016.AntGC1             : int  NA NA NA NA NA NA NA NA NA NA ...
$ Muench.2016.AntGC3             : int  NA NA NA NA NA NA NA NA NA NA ...
$ Bruyne.1999.WT                 : int  NA NA NA NA NA NA NA NA NA NA ...
}
\description{
The weights used for the global normalization across responding units.
}
\keyword{DoOR}
\keyword{dataset}
\keyword{door_global_normalization_weights}
\keyword{responding_unit}
