% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{door_mappings}
\alias{door_mappings}
\title{door_mappings}
\format{'data.frame':	96 obs. of  20 variables:
$ receptor        : Factor w/ 95 levels "?","ab2B","ab3B",..: 1 52 89 18 35 70 60 87 64 2 ...
$ sensillum       : Factor w/ 27 levels "","?","ab1","ab10",..: 2 3 3 3 3 4 4 4 5 5 ...
$ OSN             : Factor w/ 53 levels "","?","ab10A",..: 2 5 6 7 8 3 4 4 9 10 ...
$ glomerulus      : Factor w/ 57 levels "","?","D","DA1",..: 37 19 31 28 13 25 17 17 22 23 ...
$ co.receptor     : Factor w/ 9 levels "","-","?","Ir25a",..: 3 8 8 1 8 8 8 8 8 8 ...
$ coexpressing    : Factor w/ 27 levels "","Gr10a","Gr21a+Gr63a",..: 1 1 1 3 2 1 26 16 1 21 ...
$ related1        : Factor w/ 29 levels "","ac1","ac1A",..: 1 1 1 1 9 1 28 21 1 25 ...
$ related2        : Factor w/ 19 levels "","ab2B","ab3B",..: 1 1 1 1 1 1 1 1 1 13 ...
$ related3        : Factor w/ 8 levels "","ab2B","ac1BC",..: 1 1 1 1 1 1 1 1 1 1 ...
$ related4        : Factor w/ 5 levels "","ab5B","ac3A",..: 1 1 1 1 1 1 1 1 1 1 ...
$ related5        : Factor w/ 3 levels "","Ir75a","Ir75d": 1 1 1 1 1 1 1 1 1 1 ...
$ related6        : Factor w/ 3 levels "","Ir75d","Ir92a": 1 1 1 1 1 1 1 1 1 1 ...
$ Ors             : Factor w/ 50 levels "","?","Gr21a+Gr63a",..: 2 24 48 3 14 35 30 30 32 44 ...
$ sensillum.type  : Factor w/ 7 levels "","antennal basiconic",..: 1 2 2 2 2 2 2 2 2 2 ...
$ adult           : logi  NA TRUE TRUE TRUE TRUE TRUE ...
$ larva           : logi  NA TRUE NA TRUE NA NA ...
$ dataset.existing: logi  FALSE TRUE TRUE TRUE TRUE TRUE ...
$ comment         : Factor w/ 21 levels "","formerly VM6 (Grabe et al. 2014",..: 1 1 1 1 1 1 1 1 1 1 ...
$ code            : Factor w/ 52 levels "D","DA1","DA2",..: 34 17 28 25 11 22 NA 15 20 21 ...
$ code.OSN        : Factor w/ 54 levels "","ab10A","ab10B",..: NA 4 5 6 7 2 NA 3 8 9 ...}
\description{
The mappings of responding units in DoOR to morphological structures like seonsory neurons,
sensillae, etc.
}
\keyword{DoOR}
\keyword{dataset}
\keyword{door_mappings}
\keyword{responding_unit}

