% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{door_response_matrix}
\alias{door_response_matrix}
\title{door_response_matrix}
\format{'data.frame':	693 obs. of  78 variables:
$ ac1A       : num  0.0627 0.0226 0.4768 0.061 NA ...
$ ac1B       : num  0.0697 NA NA 0.1133 NA ...
$ ac2A       : num  0.056 0.0868 0.0496 0.8839 NA ...
$ ac2B       : num  0.0709 NA NA 0.0443 NA ...
$ ac3A       : num  0.0686 0.2225 0.0719 0.1899 0.1655 ...
$ ac3B       : num  0.0447 0.0612 0.3783 0.4122 0.0628 ...
$ ac4        : num  0.0927 0.1048 0.151 0.0921 0.0867 ...
$ Or1a       : num  0 0.0249 NA NA NA ...
$ Or2a       : num  0.0485 NA 0.0566 0.0638 0.0519 ...
$ Or7a       : num  0.0253 0.0944 0.0675 0.0504 0.0189 ...
$ Or9a       : num  0.0643 0.0688 0.1084 0.0979 0.1399 ...
$ Or10a      : num  0.0517 0.0567 0.0539 0.0567 0.0464 ...
$ Or13a      : num  0.0798 0.1977 0.1234 0.1535 NA ...
$ Or19a      : num  0.1001 NA 0.0972 0.0807 0.086 ...
$ Or22a      : num  0.0845 0.368 0.185 0.1751 0.0899 ...
$ Or22b      : num  NA NA NA NA NA NA NA NA NA NA ...
$ Or22c      : logi  NA NA NA NA NA NA ...
$ Or23a      : num  0.01932 NA 0.01502 0.01502 0.00429 ...
$ Or24a      : logi  NA NA NA NA NA NA ...
$ Or30a      : num  0.0124 0.0435 NA NA NA ...
$ Or33a      : num  0 0.0171 NA NA NA ...
$ Or33b      : num  0.0628 NA 0.00813 0.02421 0.0142 ...
$ Or33c      : num  0.0456 NA NA NA NA ...
$ Or35a      : num  0.0816 NA 0.0732 0.0599 0.0997 ...
$ Or42a      : num  0.082 0.123 0.031 0.175 NA ...
$ Or42b      : num  0.0516 0.2246 0.1126 0.2586 NA ...
$ Or43a      : num  0.0953 NA 0.0907 0.0322 0.0483 ...
$ Or43b      : num  0.0149 0.0654 0.0196 0.0493 0.0532 ...
$ Or45a      : num  0.00866 0.25516 NA NA NA ...
$ Or45b      : num  0.0248 0.028 NA NA NA ...
$ Or46a      : num  0.0981 NA NA NA NA ...
$ Or47a      : num  0.0519 NA 0.0814 0.0221 0.0237 ...
$ Or47b      : num  0.198 NA 0.178 0.133 0.158 ...
$ Or49a      : num  0 0.012 NA NA NA ...
$ Or49b      : num  0.048 0.0791 0.0561 0.0641 0.0552 ...
$ Or59a      : num  0.00684 NA NA NA NA ...
$ Or59b      : num  0.0237 0.0427 0.0115 0.0297 0.0276 ...
$ Or59c      : num  0.0358 0.0369 0.0348 0.0644 NA ...
$ Or65a      : num  0.0595 NA 0.0693 0.0527 0.0561 ...
$ Or67a      : num  0.02455 NA 0.04155 0.00716 0.00239 ...
$ Or67b      : num  0.0509 0.1162 NA NA NA ...
$ Or67c      : num  0.0229 0.0258 0.0259 0.0421 0.0389 ...
$ Or67d      : logi  NA NA NA NA NA NA ...
$ Or71a      : num  0.0305 0.0212 0.0217 0.0362 NA ...
$ Or74a      : num  0.0128 0.2488 NA NA NA ...
$ Or82a      : num  0.0266 0.0618 0.0301 0.0471 0.0206 ...
$ Or85a      : num  0.0457 NA 0.0332 0.021 0.0244 ...
$ Or85b      : num  0.065 0.22 0.054 0.167 0.101 ...
$ Or85c      : num  0.0293 0.101 NA NA NA ...
$ Or85d      : num  0.0472 0.0935 0.0448 0.0815 NA ...
$ Or85e      : num  0 NA NA NA NA NA NA NA NA NA ...
$ Or85f      : num  0.0339 NA 0.0866 0.061 0.0839 ...
$ Or88a      : num  0.0563 NA 0.0563 0.0438 0.0563 ...
$ Or92a      : num  0.0341 0.0244 0.0109 0.0172 NA ...
$ Or94a      : num  0.0434 0.2 NA NA NA ...
$ Or94b      : num  0.0036 0.0403 NA NA NA ...
$ Or98a      : num  0.0152 0.148 0.0682 0.0727 0.0787 ...
$ Gr21a.Gr63a: num  0.141 NA NA NA NA ...
$ ab2B       : num  0.00697 0.00799 0.00608 0.00655 NA ...
$ ab4B       : num  0.0649 NA NA 0.3068 0.0251 ...
$ ab5B       : num  0.0248 0.0923 0.0214 0.1222 NA ...
$ pb2A       : logi  NA NA NA NA NA NA ...
$ Or69a      : num  0 NA NA NA NA NA NA NA NA NA ...
$ ac1        : num  0.186 0.257 0.698 0.244 0.423 ...
$ ac2        : num  0.171 0.267 0.176 0.913 0.687 ...
$ ac3_noOr35a: num  0.0965 0.1568 0.2105 0.1258 0.0986 ...
$ Ir31a      : num  0.11669 0.13877 0.00169 0.13945 0.07324 ...
$ Ir41a      : num  0.163 0.329 0 0.491 0.414 ...
$ Ir75a      : num  0.0844 0.3045 0.1213 0.2744 0.1962 ...
$ Ir75d      : num  0.0269 0.0144 0.0324 0.0265 0.0288 ...
$ Ir76a      : num  0.0448 0.0339 0.0579 0.0286 0.0345 ...
$ Ir84a      : num  0.17465 0.20146 0.00433 0.16241 0.16627 ...
$ Ir92a      : num  0.0756 0 0.0166 0.0308 0.0399 ...
$ Ir64a.DC4  : num  0 0.855 0.0284 0.3262 0.2932 ...
$ Ir64a.DP1m : num  0.016 0.181 0.0215 0.1477 0.1019 ...
$ ac1BC      : num  0.0207 0.0252 0.023 0.0252 NA ...
$ ac2BC      : num  0.0205 0.0227 0.0244 0.0307 NA ...
$ Or83c      : num  0.0783 NA NA NA NA ...}
\description{
The actual DoOR response matrix containing the consensus data as produced by
create_door_database(). Data are globally normalized across responding units.
}
\keyword{DoOR}
\keyword{dataset}
\keyword{door_response_matrix}
\keyword{responding_unit}

