% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{door_response_matrix_non_normalized}
\alias{door_response_matrix_non_normalized}
\title{door_response_matrix_non_normalized}
\format{'data.frame':	693 obs. of  78 variables:
$ ac1A       : num  0.1144 0.0413 0.8698 0.1112 NA ...
$ ac1B       : num  0.421 NA NA 0.684 NA ...
$ ac2A       : num  0.0634 0.0982 0.0562 1 NA ...
$ ac2B       : num  0.348 NA NA 0.217 NA ...
$ ac3A       : num  0.143 0.462 0.149 0.394 0.344 ...
$ ac3B       : num  0.0574 0.0786 0.486 0.5294 0.0807 ...
$ ac4        : num  0.218 0.246 0.355 0.217 0.204 ...
$ Or1a       : num  0 0.194 NA NA NA ...
$ Or2a       : num  0.204 NA 0.238 0.269 0.219 ...
$ Or7a       : num  0.0298 0.1111 0.0794 0.0593 0.0222 ...
$ Or9a       : num  0.0988 0.1056 0.1664 0.1502 0.2147 ...
$ Or10a      : num  0.0642 0.0705 0.067 0.0705 0.0576 ...
$ Or13a      : num  0.0919 0.2275 0.142 0.1766 NA ...
$ Or19a      : num  0.1143 NA 0.1111 0.0921 0.0982 ...
$ Or22a      : num  0.0975 0.4247 0.2135 0.2021 0.1038 ...
$ Or22b      : num  NA NA NA NA NA NA NA NA NA NA ...
$ Or22c      : logi  NA NA NA NA NA NA ...
$ Or23a      : num  0.0723 NA 0.0562 0.0562 0.0161 ...
$ Or24a      : logi  NA NA NA NA NA NA ...
$ Or30a      : num  0.0171 0.0598 NA NA NA ...
$ Or33a      : num  0 0.267 NA NA NA ...
$ Or33b      : num  0.2526 NA 0.0327 0.0974 0.0571 ...
$ Or33c      : num  0.101 NA NA NA NA ...
$ Or35a      : num  0.1004 NA 0.0902 0.0738 0.1228 ...
$ Or42a      : num  0.0914 0.137 0.0346 0.1947 NA ...
$ Or42b      : num  0.0574 0.2498 0.1252 0.2877 NA ...
$ Or43a      : num  0.1744 NA 0.166 0.0589 0.0884 ...
$ Or43b      : num  0.02 0.0878 0.0264 0.0661 0.0714 ...
$ Or45a      : num  0.01 0.294 NA NA NA ...
$ Or45b      : num  0.0259 0.0293 NA NA NA ...
$ Or46a      : num  0.163 NA NA NA NA ...
$ Or47a      : num  0.0598 NA 0.0937 0.0254 0.0273 ...
$ Or47b      : num  0.641 NA 0.575 0.431 0.51 ...
$ Or49a      : num  0 0.0781 NA NA NA ...
$ Or49b      : num  0.0668 0.1101 0.0781 0.0892 0.0768 ...
$ Or59a      : num  0.014 NA NA NA NA ...
$ Or59b      : num  0.0312 0.0562 0.0152 0.039 0.0363 ...
$ Or59c      : num  0.1025 0.1055 0.0996 0.1844 NA ...
$ Or65a      : num  0.26 NA 0.302 0.23 0.245 ...
$ Or67a      : num  0.03305 NA 0.05594 0.00965 0.00322 ...
$ Or67b      : num  0.0512 0.1169 NA NA NA ...
$ Or67c      : num  0.0453 0.0511 0.0513 0.0835 0.0771 ...
$ Or67d      : logi  NA NA NA NA NA NA ...
$ Or71a      : num  0.0372 0.026 0.0265 0.0442 NA ...
$ Or74a      : num  0.0246 0.4775 NA NA NA ...
$ Or82a      : num  0.0363 0.0843 0.041 0.0643 0.0282 ...
$ Or85a      : num  0.0668 NA 0.0485 0.0306 0.0357 ...
$ Or85b      : num  0.077 0.2605 0.0641 0.1984 0.1201 ...
$ Or85c      : num  0.0335 0.1154 NA NA NA ...
$ Or85d      : num  0.0835 0.1657 0.0794 0.1444 NA ...
$ Or85e      : num  0 NA NA NA NA NA NA NA NA NA ...
$ Or85f      : num  0.118 NA 0.301 0.212 0.292 ...
$ Or88a      : num  0.263 NA 0.262 0.204 0.262 ...
$ Or92a      : num  0.0472 0.0337 0.0151 0.0238 NA ...
$ Or94a      : num  0.047 0.217 NA NA NA ...
$ Or94b      : num  0.00785 0.08806 NA NA NA ...
$ Or98a      : num  0.0191 0.1863 0.0858 0.0915 0.0991 ...
$ Gr21a.Gr63a: num  0.141 NA NA NA NA ...
$ ab2B       : num  0.0143 0.0165 0.0125 0.0135 NA ...
$ ab4B       : num  0.1017 NA NA 0.4812 0.0394 ...
$ ab5B       : num  0.0262 0.0972 0.0225 0.1286 NA ...
$ pb2A       : logi  NA NA NA NA NA NA ...
$ Or69a      : num  0 NA NA NA NA NA NA NA NA NA ...
$ ac1        : num  0.266 0.368 1 0.349 0.605 ...
$ ac2        : num  0.171 0.267 0.176 0.913 0.687 ...
$ ac3_noOr35a: num  0.149 0.241 0.324 0.194 0.152 ...
$ Ir31a      : num  0.5038 0.5991 0.0073 0.6021 0.3162 ...
$ Ir41a      : num  0.331 0.67 0 1 0.843 ...
$ Ir75a      : num  0.131 0.473 0.188 0.426 0.305 ...
$ Ir75d      : num  0.0269 0.0144 0.0324 0.0265 0.0288 ...
$ Ir76a      : num  0.774 0.585 1 0.494 0.596 ...
$ Ir84a      : num  0.3837 0.4426 0.0095 0.3568 0.3653 ...
$ Ir92a      : num  0.2147 0 0.0472 0.0874 0.1133 ...
$ Ir64a.DC4  : num  0 0.855 0.0284 0.3262 0.2932 ...
$ Ir64a.DP1m : num  0.0261 0.2957 0.0351 0.2414 0.1665 ...
$ ac1BC      : num  0.364 0.443 0.403 0.443 NA ...
$ ac2BC      : num  0.581 0.645 0.694 0.871 NA ...
$ Or83c      : num  0.0783 NA NA NA NA ...}
\description{
The actual DoOR response matrix containing the consensus data as produced by
create_door_database(). Data are not normalized.
}
\keyword{DoOR}
\keyword{dataset}
\keyword{door_response_matrix_non_normalized}
\keyword{responding_unit}

