% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{door_response_range}
\alias{door_response_range}
\title{door_response_range}
\format{'data.frame':	42 obs. of  4 variables:
$ study  : Factor w/ 41 levels "Bruyne.1999.WT",..: 2 3 4 5 6 7 8 9 10 11 ...
$ min    : num  0 -1.2 -5.25 0 0 -10 -16 -20.4 -34 -0.15 ...
$ max    : num  203 235 269 276 114 ...
$ n_odors: int  50 13 13 19 12 474 43 12 101 22 ...}
\description{
The response ranges of the respective data sets.
}
\keyword{DoOR}
\keyword{dataset}
\keyword{door_response_range}
\keyword{responding_unit}

