% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{door_excluded_data}
\alias{door_excluded_data}
\title{door_excluded_data}
\format{A data frame with 78 observations on 2 variables.}
\description{
datasets that were excluded during the merging process
}
\keyword{datasets}

