% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{glo.dist}
\alias{glo.dist}
\title{a matrix of distance coefficients between glomeruli}
\format{A data frame with 49 observations on the following 49 variables.
\describe{ \item{list("D")}{a logical vector} \item{list("DA1")}{a numeric
vector} \item{list("DA2")}{a numeric vector} \item{list("DA3")}{a numeric
vector} \item{list("DA4l")}{a numeric vector} \item{list("DA4m")}{a numeric
vector} \item{list("DC1")}{a numeric vector} \item{list("DC2")}{a numeric
vector} \item{list("DC3")}{a numeric vector} \item{list("DL1")}{a numeric
vector} \item{list("DL2")}{a numeric vector} \item{list("DL3")}{a numeric
vector} \item{list("DL4")}{a numeric vector} \item{list("DL5")}{a numeric
vector} \item{list("DM1")}{a numeric vector} \item{list("DM2")}{a numeric
vector} \item{list("DM3")}{a numeric vector} \item{list("DM4")}{a numeric
vector} \item{list("DM5")}{a numeric vector} \item{list("DM6")}{a numeric
vector} \item{list("DP1l")}{a numeric vector} \item{list("DP1m")}{a numeric
vector} \item{list("extra.glomerulus1")}{a numeric vector}
\item{list("extra.glomerulus2")}{a numeric vector} \item{list("V")}{a
numeric vector} \item{list("VA1d")}{a numeric vector} \item{list("VA1lm")}{a
numeric vector} \item{list("VA1v")}{a numeric vector} \item{list("VA2")}{a
numeric vector} \item{list("VA3")}{a numeric vector} \item{list("VA4")}{a
numeric vector} \item{list("VA5")}{a numeric vector} \item{list("VA6")}{a
numeric vector} \item{list("VA7l")}{a numeric vector} \item{list("VA7m")}{a
numeric vector} \item{list("VC1")}{a numeric vector} \item{list("VC2")}{a
numeric vector} \item{list("VC3")}{a numeric vector} \item{list("VC4")}{a
numeric vector} \item{list("VL1")}{a numeric vector} \item{list("VL2a")}{a
numeric vector} \item{list("VL2p")}{a numeric vector} \item{list("VM1")}{a
numeric vector} \item{list("VM2")}{a numeric vector} \item{list("VM3")}{a
numeric vector} \item{list("VM4")}{a numeric vector} \item{list("VM5d")}{a
numeric vector} \item{list("VM6")}{a numeric vector} \item{list("VM7")}{a
numeric vector} }}
\description{
a matrix of distance coefficients between glomeruli
}
\details{
The odorant distance coefficients between glomeruli were computed by using
the VRML model from http://flybrain.neurobio.arizona.edu
}
\examples{
data(glo.dist)
}
\references{
\itemize{ \item \ http://flybrain.neurobio.arizona.edu \item
\\Laissue, P. P.; Reiter, C.; Hiesinger, P. R.; Halter, S.; Fischbach, K. F.
& Stocker, R. F., 1999, \emph{Three-dimensional reconstruction of the
antennal lobe in Drosophila melanogaster.}, J Comp Neurol,, 405:543-552 }
}
\keyword{datasets}

