% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Ir64a.DC4}
\alias{Ir64a.DC4}
\title{Ir64a.DC4}
\format{A data frame with 693 observations on 6 variables.}
\description{
DoOR response data
}
\details{
measurements from the Ir64a expressing OSNs that project to DC4
}
\keyword{datasets}

