% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Ir76b}
\alias{Ir76b}
\title{Ir76b}
\format{A data frame with 251 observations on the following 4 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector} \item{list("CAS")}{a factor with
odorant CAS numbers} }}
\description{
Response profile of Ir76b
}
\details{
Ir76b \describe{ \item{list("Sequence location")}{ 3L:20,101,884..20,104,537
[-] \link[=reference]{(Tweedie et al., 2009)} }\item{:}{
3L:20,101,884..20,104,537 [-] \link[=reference]{(Tweedie et al., 2009)} }
\item{list("Sensilla")}{ antennal coeloconic sensilla (antenna)
\link[=reference]{ (Benton at. al., 2009) } }\item{:}{ antennal coeloconic
sensilla (antenna) \link[=reference]{ (Benton at. al., 2009) } }
\item{list("Neuron")}{ ac1, ac2, ac3, ac4 \link[=reference]{ (Benton at.
al., 2009) } }\item{:}{ ac1, ac2, ac3, ac4 \link[=reference]{ (Benton at.
al., 2009) } } \item{list("Coexpression")}{ with Ir92a in ac1, with Or35a in
ac3, with Ir76a in ac4, express alone in ac2 \link[=reference]{ (Benton at.
al., 2009) }. }\item{:}{ with Ir92a in ac1, with Or35a in ac3, with Ir76a in
ac4, express alone in ac2 \link[=reference]{ (Benton at. al., 2009) }. }

\item{list("Comment")}{ \enumerate{ \item IRs functions as chemosenor that
does not express ORs or Or83b \link[=reference]{ (Benton at. al., 2009) }.
\item The neurons ac2 that express Ir76b alone are sensitive to
1,4-diaminobutane \link[=reference]{ (Benton at. al., 2009) }.  } }\item{:}{
\enumerate{ \item IRs functions as chemosenor that does not express ORs or
Or83b \link[=reference]{ (Benton at. al., 2009) }.  \item The neurons ac2
that express Ir76b alone are sensitive to 1,4-diaminobutane
\link[=reference]{ (Benton at. al., 2009) }.  } } }
}
\examples{
data(Ir76b)
}
\references{
\itemize{ \item \link[=reference]{ (Benton at. al., 2009) } : R.
Benton, K. S. Vannice, C. Gomez-Diaz, L. B. Vosshall, \emph{Variant
Ionotropic Glutamate Receptors as Chemosensory Receptors in Drosophila},
2009, cell, 136(1):149-162 }
}
\keyword{datasets}

