% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{load_door_data}
\alias{loadData}
\alias{load_door_data}
\title{load_door_data}
\usage{
load_door_data()
}
\value{
attaches all DoOR data to the main workspace
}
\description{
load all DoOR.data
}
\examples{

load_door_data()

}
\author{
Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}

