% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{loadData}
\usage{
loadData()
}
\value{
attaches all DoOR data to the main workspace
}
\description{
load all DoOR.data
}
\examples{

loadData()

}
\author{
Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}

