% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{OGN}
\alias{OGN}
\title{Mapping of Receptors}
\format{A data frame with 84 observations on the following 12 variables.
\describe{ \item{list("Receptor")}{a factor with receptor or neuron names}
\item{list("Coexpress")}{a factor with coespression}
\item{list("Coexpressing1")}{a factor with coespression1}
\item{list("Coexpressing2")}{a factor with coespression2}
\item{list("Coexpressing3")}{a factor with coespression3}
\item{list("Ors")}{a factor with receptor names} \item{list("Sensillium")}{a
factor with sensillum type} \item{list("ORN")}{a factor with neuron names}
\item{list("Glomerulus")}{a factor with glomerulus names}
\item{list("Code")}{a numeric vector} \item{list("x")}{a numeric vector}
\item{list("y")}{a numeric vector} }}
\description{
The map of Receptor to sensillium, receptor neuron and glomerulus
}
\details{
The data frame is used for showing functional antennal lobe.
}
\examples{
data(OGN)
}
\keyword{datasets}

