% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Or13a}
\alias{Or13a}
\title{Or13a}
\format{'data.frame':	693 obs. of  12 variables:
$ Class            : Factor w/ 17 levels "acid","acids",..: NA 13 5 5 5 5 5 5 5 5 ...
$ Name             : Factor w/ 690 levels "11-cis vaccenyl acetate",..: 634 675 240 613 283 239 363 436 458 341 ...
$ InChIKey         : Factor w/ 693 levels "ACCRBMDJCPPJDX-UHFFFAOYSA-N",..: 482 612 549 252 548 418 196 577 41 462 ...
$ CID              : Factor w/ 687 levels "1001","10050",..: 686 680 139 15 220 189 483 610 564 468 ...
$ CAS              : Factor w/ 677 levels "1001-45-2","10032-13-0",..: 676 591 212 100 379 586 231 114 62 200 ...
$ Schmuker.2007.TR : int  14 NA NA 17 NA 4 8 NA NA NA ...
$ Nissler.2007.EC50: num  0 NA NA NA NA NA NA NA NA NA ...
$ Nissler.2007.nmr : num  0 NA NA NA NA NA NA NA NA NA ...
$ Kreher.2008.EN   : int  5 NA NA NA NA NA NA NA NA NA ...
$ Bruyne.2001.WT   : num  14 NA NA NA NA NA NA NA NA NA ...
$ Montague.2011.EN : int  17 15 NA NA NA NA NA NA NA NA ...
$ Marshall.2010.WT : num  0 27.33 2.67 12 NA ...}
\description{
DoOR response data for responding unit Or13a. Please find detailed
information on the respective sources of the data in door_dataset_info.
}
\keyword{DoOR}
\keyword{Or13a}
\keyword{dataset}
\keyword{responding_unit}
