% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Or22b}
\alias{Or22b}
\title{Or22b}
\format{A data frame with 251 observations on the following 5 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector with compound ID }
\item{list("CAS")}{a factor with odorant CAS numbers}
\item{list("Dobritsa.2003.EN")}{a numeric vector; electrophysiological
recording in empty neuron from Dobritsa et al. 2003
\link[=reference]{(Dobritsa et al., 2003)} } }}
\description{
Response profile of Or22b
}
\details{
Or22b \describe{ \item{list("Sequence location")}{ 2L:2,054,589..2,057,967
[+] \link[=reference]{(Tweedie et al., 2009)} }\item{:}{
2L:2,054,589..2,057,967 [+] \link[=reference]{(Tweedie et al., 2009)} }
\item{list("Expression")}{ in adult \link[=reference]{(Dobritsa et al.,
2003)} }\item{:}{ in adult \link[=reference]{(Dobritsa et al., 2003)} }
\item{list("Sensilla")}{ large basiconic sensilla (antenna)
\link[=reference]{(Dobritsa et al., 2003)}. }\item{:}{ large basiconic
sensilla (antenna) \link[=reference]{(Dobritsa et al., 2003)}. }
\item{list("Neuron")}{ ab3A \link[=reference]{(Dobritsa et al., 2003)}.
}\item{:}{ ab3A \link[=reference]{(Dobritsa et al., 2003)}. }
\item{list("Coexpression")}{ with Or22a \link[=reference]{(Dobritsa et al.,
2003)} }\item{:}{ with Or22a \link[=reference]{(Dobritsa et al., 2003)} } }
}
\examples{
data(Or22b)
}
\references{
\itemize{ \item \link[=reference]{(Dobritsa et al., 2003)} :
Dobritsa A. A., van der Goes van Naters W., Warr C.G., Steinbrecht R. A.,
Carlson J.R., 2003, \emph{Integrating the molecular and cellular basis of
odor coding in the Drosophila antenna}, Neuron, 37:827-841 }
}
\keyword{datasets}

