% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Or59a}
\alias{Or59a}
\title{Or59a}
\format{A data frame with 251 observations on the following 6 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector with compound ID }
\item{list("CAS")}{a factor with odorant CAS numbers}
\item{list("Kreher.2005.EN")}{a numeric vector; electrophysiological
recording in empty neuron from Kreher et al. 2005 \link[=reference]{(Kreher
et al., 2005)} } \item{list("Kreher.2008.EN")}{a numeric vector;
electrophysiological recording in empty neuron from Kreher et al. 2008
\link[=reference]{(Kreher et al., 2008)} } }}
\description{
Response profile of Or59a
}
\details{
Or59a \describe{ \item{list("Sequence location")}{ 2R:19,327,392..19,328,588
[-] \link[=reference]{(Tweedie et al., 2009)} }\item{:}{
2R:19,327,392..19,328,588 [-] \link[=reference]{(Tweedie et al., 2009)} }
\item{list("Expression")}{ larval specific \link[=reference]{(Couto et al.,
2005)}, \link[=reference]{(Fishilevich et al., 2005)}; \cr Larval specific
odor receptor 59a was detected by in situ hybridization
\link[=reference]{(Fishilevich et al., 2005)}.  Or59a was not detected in
adult olfactory organs by in situ hybridization analysis
\link[=reference]{(Vosshall et al., 2000)}.}\item{:}{ larval specific
\link[=reference]{(Couto et al., 2005)}, \link[=reference]{(Fishilevich et
al., 2005)}; \cr Larval specific odor receptor 59a was detected by in situ
hybridization \link[=reference]{(Fishilevich et al., 2005)}.  Or59a was not
detected in adult olfactory organs by in situ hybridization analysis
\link[=reference]{(Vosshall et al., 2000)}.} \item{list("Comment")}{
Supplemental odor response information: Or59a responded strongly to anisole
(CAS: 100-66-3) and methyl eugenol (CAS: 93-15-2). However, anisole elicited
a response with fast onset kinetics, whereas methyl eugenol elicited a
response with long lasting kinetics \link[=reference]{(van der Goes van
Naters and Carlson, 2007)}. }\item{:}{ Supplemental odor response
information: Or59a responded strongly to anisole (CAS: 100-66-3) and methyl
eugenol (CAS: 93-15-2). However, anisole elicited a response with fast onset
kinetics, whereas methyl eugenol elicited a response with long lasting
kinetics \link[=reference]{(van der Goes van Naters and Carlson, 2007)}. } }
}
\examples{
data(Or59a)
}
\references{
\itemize{ \item \link[=reference]{(Kreher et al., 2005)} :
Kreher S. A., Kwon J. Y. And Carlson J. R., 2005, \emph{The Molecular Basis
of Odor Coding in the Drosophila Larva}, Neuron, 46;445-456 \item
\link[=reference]{(Kreher et al., 2008)} : Kreher S. A., Mathew D., Kim J.,
Carlson J. R., 2008, \emph{Translation of Sensory Input into Behavioral
Output via an Olfactory System}, Neuron, 59:110-124 }
}
\keyword{datasets}

