% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Or59c}
\alias{Or59c}
\title{Or59c}
\format{A data frame with 251 observations on the following 6 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector with compound ID }
\item{list("CAS")}{a factor with odorant CAS numbers}
\item{list("Goldman.2005.WT")}{a numeric vector; electrophysiological
recording in wild type from Goldman et al. 2005 \link[=reference]{(Goldman
et al., 2005)} } \item{list("Bruyne.1999.WT")}{a numeric vector;
electrophysiological recording in wild type from de Bruyne et al. 1999
\link[=reference]{(de Bruyne et al., 1999)} } }}
\description{
Response profile of Or59c
}
\details{
Or59c \describe{ \item{list("Sequence location")}{ 2R:19,359,836..19,361,126
[-] \link[=reference]{(Tweedie et al., 2009)} }\item{:}{
2R:19,359,836..19,361,126 [-] \link[=reference]{(Tweedie et al., 2009)} }
\item{list("Expression")}{ in adult \link[=reference]{(Goldman et al.,
2005)}, \link[=reference]{(Couto et al., 2005)}, \link[=reference]{(Vosshall
et al., 2000)} }\item{:}{ in adult \link[=reference]{(Goldman et al.,
2005)}, \link[=reference]{(Couto et al., 2005)}, \link[=reference]{(Vosshall
et al., 2000)} } \item{list("Sensilla")}{ thin basiconic sensilla (palp)
\link[=reference]{(Goldman et al., 2005)} }\item{:}{ thin basiconic sensilla
(palp) \link[=reference]{(Goldman et al., 2005)} } \item{list("Neuron")}{
pb3A \link[=reference]{(Goldman et al., 2005)} }\item{:}{ pb3A
\link[=reference]{(Goldman et al., 2005)} } \item{list("Coexpression")}{ no
recorded }\item{:}{ no recorded } \item{list("Glomerulus")}{ 1 (VM7)
\link[=reference]{(Couto et al., 2005)} \cr Or59c was observed label two
glomeruli \link[=reference]{(Couto et al., 2005)}. One of these glomeruli:
(1) is likely to be target by endogenous receptor neuron, the other
glomerulus (VM7) is likely due to ectopic expression of receptor
\link[=reference]{(Couto et al., 2005)}. Reported by Fishilevich, Or59c was
showed ectopic expression \link[=reference]{(Fishilevich and Vosshall,
2005)}. }\item{:}{ 1 (VM7) \link[=reference]{(Couto et al., 2005)} \cr Or59c
was observed label two glomeruli \link[=reference]{(Couto et al., 2005)}.
One of these glomeruli: (1) is likely to be target by endogenous receptor
neuron, the other glomerulus (VM7) is likely due to ectopic expression of
receptor \link[=reference]{(Couto et al., 2005)}. Reported by Fishilevich,
Or59c was showed ectopic expression \link[=reference]{(Fishilevich and
Vosshall, 2005)}. } }
}
\examples{
data(Or59c)
}
\references{
\itemize{ \item \link[=reference]{(Goldman et al., 2005)} :
Goldman A. L., Van der Goes van Naters W., Lessing D, Warr C. G., Carlson J.
R., 2005, \emph{Coexpression of Two Functional Odor Receptors in one
Neuron}, Neuron, 45:661-666 \item \link[=reference]{(de Bruyne et al.,
1999)} : de Bruyne M., Clyne P. J., Carlson J. R., 1999, \emph{Odor coding
in a model olfactory organ: The Drosophila maxillary palp}, J. Neurosci.,
19:4520-4532 }
}
\keyword{datasets}

