% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Or67b}
\alias{Or67b}
\title{Or67b}
\format{A data frame with 251 observations on the following 7 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector with compound ID }
\item{list("CAS")}{a factor with odorant CAS numbers}
\item{list("Kreher.2005.EN")}{a numeric vector; electrophysiological
recording in empty neuron from Kreher et al. 2005 \link[=reference]{(Kreher
et al., 2005)} } \item{list("Kreher.2008.EN")}{a numeric vector;
electrophysiological recording in empty neuron from Kreher et al. 2008
\link[=reference]{(Kreher et al., 2008)} } \item{list("Galizia.2009.nmr")}{a
numeric vector; normalized mean values of calcium imaging recording. } }}
\description{
Response profile of Or67b
}
\details{
Or67b \describe{ \item{list("Sequence location")}{ 3L:9,620,512..9,622,293
[-] \link[=reference]{(Tweedie et al., 2009)} }\item{:}{
3L:9,620,512..9,622,293 [-] \link[=reference]{(Tweedie et al., 2009)} }
\item{list("Expression")}{ in adult \link[=reference]{(Couto et al., 2005)}
and larvae \link[=reference]{(Kreher et al., 2005)},
\link[=reference]{(Fishilevich et al., 2005)} }\item{:}{ in adult
\link[=reference]{(Couto et al., 2005)} and larvae \link[=reference]{(Kreher
et al., 2005)}, \link[=reference]{(Fishilevich et al., 2005)} }
\item{list("Sensilla")}{ small basiconic sensilla (antenna)
\link[=reference]{(Couto et al., 2005)} }\item{:}{ small basiconic sensilla
(antenna) \link[=reference]{(Couto et al., 2005)} } \item{list("Neuron")}{
ab9 \link[=reference]{(Couto et al., 2005)} }\item{:}{ ab9
\link[=reference]{(Couto et al., 2005)} } \item{list("Coexpression")}{ no
recorded }\item{:}{ no recorded } \item{list("Glomerulus")}{ VA3
\link[=reference]{(Couto et al., 2005)} }\item{:}{ VA3
\link[=reference]{(Couto et al., 2005)} } }
}
\examples{
data(Or67b)
}
\references{
\itemize{ \item \link[=reference]{(Kreher et al., 2005)} :
Kreher S. A., Kwon J. Y. And Carlson J. R., 2005, \emph{The Molecular Basis
of Odor Coding in the Drosophila Larva}, Neuron, 46;445-456 \item
\link[=reference]{(Kreher et al., 2008)} : Kreher S. A., Mathew D., Kim J.,
Carlson J. R., 2008, \emph{Translation of Sensory Input into Behavioral
Output via an Olfactory System}, Neuron, 59:110-124 }
}
\keyword{datasets}

