% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Or67c}
\alias{Or67c}
\title{Or67c}
\format{A data frame with 251 observations on the following 8 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector with compound ID }
\item{list("CAS")}{a factor with odorant CAS numbers}
\item{list("Hallem.2006.EN")}{a numeric vector; electrophysiological
recording in empty neuron from Hallem et al. 2006 \link[=reference]{(Hallem
and Carlson, 2006)} } \item{list("Hallem.2004.EN")}{a numeric vector;
electrophysiological recording in empty neuron from Hallem et al. 2004
\link[=reference]{(Hallem et al., 2004)} } \item{list("Bruyne.2001.WT")}{a
numeric vector; electrophysiological recording in wild type from de Bruyne
et al. 2001 \link[=reference]{(de Bruyne et al., 2001)} }
\item{list("Bruyne.2001.RR")}{a numeric vector; electrophysiological
recording data of response range from de Bruyne et al. 2001
\link[=reference]{(de Bruyne et al., 2001)} } }}
\description{
Response profile of Or67c
}
\details{
Or67c \describe{ \item{list("Sequence location")}{ 2L:15,622,128..15,623,558
[-] \link[=reference]{(Tweedie et al., 2009)} }\item{:}{
2L:15,622,128..15,623,558 [-] \link[=reference]{(Tweedie et al., 2009)} }
\item{list("Expression")}{ in adult \link[=reference]{(Couto et al., 2005)},
\link[=reference]{[23]} and larvae \link[=reference]{(Couto et al., 2005)}
}\item{:}{ in adult \link[=reference]{(Couto et al., 2005)},
\link[=reference]{[23]} and larvae \link[=reference]{(Couto et al., 2005)} }
\item{list("Sensilla")}{ small basiconic sensilla (antenna)
\link[=reference]{(Couto et al., 2005)} }\item{:}{ small basiconic sensilla
(antenna) \link[=reference]{(Couto et al., 2005)} } \item{list("Neuron")}{
ab7 \link[=reference]{(Couto et al., 2005)}; \cr ab7 houses two classes of
ORs, Or98a and Or67c \link[=reference]{(Couto et al., 2005)}. Or98a was
mapped to ab7A by comparing odor response spectrum of delta ab3A:OrX to wild
type ORN \link[=reference]{(Hallem et al., 2004)}, suggesting that Or67c is
mapped to ab7B. }\item{:}{ ab7 \link[=reference]{(Couto et al., 2005)}; \cr
ab7 houses two classes of ORs, Or98a and Or67c \link[=reference]{(Couto et
al., 2005)}. Or98a was mapped to ab7A by comparing odor response spectrum of
delta ab3A:OrX to wild type ORN \link[=reference]{(Hallem et al., 2004)},
suggesting that Or67c is mapped to ab7B. } \item{list("Coexpression")}{ no
recorded }\item{:}{ no recorded } \item{list("Glomerulus")}{ VC4
\link[=reference]{(Couto et al., 2005)}, VC3m \link[=reference]{(Laissue and
Vosshall, 2008)} }\item{:}{ VC4 \link[=reference]{(Couto et al., 2005)},
VC3m \link[=reference]{(Laissue and Vosshall, 2008)} } }
}
\examples{
data(Or67c)
}
\references{
\itemize{ \item \link[=reference]{(Hallem and Carlson, 2006)} :
Hallem E. A. & Carlson J. R., 2006, \emph{Coding of odors by a receptor
repertoire}, Cell,125:143-160 \item \link[=reference]{(Hallem et al., 2004)}
: Hallem E.A., Ho M. G. and Carlson J.R., 2004, \emph{The Molecular Basis of
Odor Coding in the Drosophila Antenna}, Cell, 117:965-979 \item
\link[=reference]{(de Bruyne et al., 2001)} : de Bruyne M., Foster K.,
Carlson J. R., 2001, \emph{Odor coding in the Drosophila antenna, Neuron},
30:537-552 }
}
\keyword{datasets}

