% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Or69a}
\alias{Or69a}
\title{Or69a}
\format{A data frame with 251 observations on the following 4 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector with compound ID }
\item{list("CAS")}{a factor with odorant CAS numbers} }}
\description{
Response profile of Or69a
}
\details{
Or69a \describe{ \item{list("Sequence location")}{3L:12,957,460..12,960,833
[-] \link[=reference]{(Tweedie et al., 2009)}
}\item{:}{3L:12,957,460..12,960,833 [-] \link[=reference]{(Tweedie et al.,
2009)} } \item{list("Expression")}{ in adult \link[=reference]{(Couto et
al., 2005)} }\item{:}{ in adult \link[=reference]{(Couto et al., 2005)} }
\item{list("Sensilla")}{ small basiconic sensilla (antenna) (antenna)
\link[=reference]{(Couto et al., 2005)}}\item{:}{ small basiconic sensilla
(antenna) (antenna) \link[=reference]{(Couto et al., 2005)}}
\item{list("Neuron")}{ ab9 \link[=reference]{(Couto et al., 2005)}
}\item{:}{ ab9 \link[=reference]{(Couto et al., 2005)} }
\item{list("Coexpression")}{ Or69aA and Or69aB coexpress in ab9,
coconvergent to glomerulus D \link[=reference]{(Couto et al., 2005)}.
}\item{:}{ Or69aA and Or69aB coexpress in ab9, coconvergent to glomerulus D
\link[=reference]{(Couto et al., 2005)}. } \item{list("Glomerulus")}{ D
\link[=reference]{(Couto et al., 2005)} }\item{:}{ D
\link[=reference]{(Couto et al., 2005)} } \item{list("Isoform")}{ Or69aA and
Or69aB \link[=reference]{(Couto et al., 2005)} }\item{:}{ Or69aA and Or69aB
\link[=reference]{(Couto et al., 2005)} } }
}
\examples{
data(Or69a)
}
\keyword{datasets}

