% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Or71a}
\alias{Or71a}
\title{Or71a}
\format{A data frame with 251 observations on the following 6 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector with compound ID }
\item{list("CAS")}{a factor with odorant CAS numbers}
\item{list("Goldman.2005.WT")}{a numeric vector; electrophysiological
recording in wild type from Goldman et al. 2005 \link[=reference]{(Goldman
et al., 2005)} } \item{list("Bruyne.1999.WT")}{a numeric vector;
electrophysiological recording in wild type from de Bruyne et al. 1999
\link[=reference]{(de Bruyne et al., 1999)} } }}
\description{
Response profile of Or71a
}
\details{
Or71a \describe{ \item{list("Sequence location")}{ 3L:15,070,303..15,071,547
[-] \link[=reference]{(Tweedie et al., 2009)} }\item{:}{
3L:15,070,303..15,071,547 [-] \link[=reference]{(Tweedie et al., 2009)} }
\item{list("Expression")}{ in adult \link[=reference]{(Vosshall et al.,
2000)} }\item{:}{ in adult \link[=reference]{(Vosshall et al., 2000)} }
\item{list("Sensilla")}{ thin basiconic sensilla (palp)
\link[=reference]{(Goldman et al., 2005)}, \link[=reference]{(Fishilevich
and Vosshall, 2005)} }\item{:}{ thin basiconic sensilla (palp)
\link[=reference]{(Goldman et al., 2005)}, \link[=reference]{(Fishilevich
and Vosshall, 2005)} } \item{list("Neuron")}{ pb1B
\link[=reference]{(Goldman et al., 2005)} }\item{:}{ pb1B
\link[=reference]{(Goldman et al., 2005)} } \item{list("Coexpression")}{ no
recorded }\item{:}{ no recorded } \item{list("Glomerulus")}{ VC2
\link[=reference]{(Couto et al., 2005)} }\item{:}{ VC2
\link[=reference]{(Couto et al., 2005)} } }
}
\examples{
data(Or71a)
}
\references{
\itemize{ \item \link[=reference]{(Goldman et al., 2005)} :
Goldman A. L., Van der Goes van Naters W., Lessing D, Warr C. G., Carlson J.
R., 2005, \emph{Coexpression of Two Functional Odor Receptors in one
Neuron}, Neuron, 45:661-666 \item \link[=reference]{(de Bruyne et al.,
1999)} : de Bruyne M., Clyne P. J., Carlson J. R., 1999, \emph{Odor coding
in a model olfactory organ: The Drosophila maxillary palp}, J. Neurosci.,
19:4520-4532 }
}
\keyword{datasets}

