% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Or82a}
\alias{Or82a}
\title{Or82a}
\format{A data frame with 251 observations on the following 10 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector with compound ID }
\item{list("CAS")}{a factor with odorant CAS numbers}
\item{list("Hallem.2006.EN")}{a numeric vector; electrophysiological
recording in empty neuron from Hallem et al. 2006 \link[=reference]{(Hallem
and Carlson, 2006)} } \item{list("Hallem.2004.EN")}{a numeric vector;
electrophysiological recording in empty neuron from Hallem et al. 2004
\link[=reference]{(Hallem et al., 2004)} } \item{list("Hallem.2004.WT")}{a
numeric vector; electrophysiological recording in wild type from Hallem et
al. 2004 \link[=reference]{(Hallem et al., 2004)} }
\item{list("Bruyne.2001.WT")}{a numeric vector; electrophysiological
recording in wild type from de Bruyne et al. 2001 \link[=reference]{(de
Bruyne et al., 2001)} } \item{list("Bruyne.2001.RR")}{a numeric vector;
electrophysiological recording data of response range from de Bruyne et al.
2001 \link[=reference]{(de Bruyne et al., 2001)} }
\item{list("Kreher.2008.EN")}{a numeric vector; electrophysiological
recording in empty neuron from Kreher et al. 2008 \link[=reference]{(Kreher
et al., 2008)} } }}
\description{
Response profile of Or82a
}
\details{
Or82a \describe{ \item{list("Sequence location")}{ 3R:82,694..84,166 [+]
\link[=reference]{(Tweedie et al., 2009)} }\item{:}{ 3R:82,694..84,166 [+]
\link[=reference]{(Tweedie et al., 2009)} } \item{list("Expression")}{ in
adult \link[=reference]{(Couto et al., 2005)}, \link[=reference]{(Vosshall
et al., 2000)}, \link[=reference]{(Fishilevich and Vosshall, 2005)} and
larvae \link[=reference]{(Fishilevich et al., 2005)} }\item{:}{ in adult
\link[=reference]{(Couto et al., 2005)}, \link[=reference]{(Vosshall et al.,
2000)}, \link[=reference]{(Fishilevich and Vosshall, 2005)} and larvae
\link[=reference]{(Fishilevich et al., 2005)} } \item{list("Sensilla")}{
small basiconic sensilla (antenna) \link[=reference]{(Couto et al., 2005)}
}\item{:}{ small basiconic sensilla (antenna) \link[=reference]{(Couto et
al., 2005)} } \item{list("Neuron")}{ ab5A \link[=reference]{(Hallem et al.,
2004)} }\item{:}{ ab5A \link[=reference]{(Hallem et al., 2004)} }
\item{list("Coexpression")}{ no recorded }\item{:}{ no recorded }
\item{list("Glomerulus")}{ VA6 \link[=reference]{(Yao et al., 2005)} \cr
Glomerulus VA6 receives input from two separate populations of OSNs
expressing different ORs: Or82a and Or67d, however, they are not coexpressed
\link[=reference]{(Fishilevich and Vosshall, 2005)}. }\item{:}{ VA6
\link[=reference]{(Yao et al., 2005)} \cr Glomerulus VA6 receives input from
two separate populations of OSNs expressing different ORs: Or82a and Or67d,
however, they are not coexpressed \link[=reference]{(Fishilevich and
Vosshall, 2005)}. } \item{list("Comment")}{ Or82a responds selectively to
geranyl acetate, a green-leaf volatile that is also a major component of
male sex pheromone \link[=reference]{(Ha and Smith, 2006)}.  }\item{:}{
Or82a responds selectively to geranyl acetate, a green-leaf volatile that is
also a major component of male sex pheromone \link[=reference]{(Ha and
Smith, 2006)}.  } }
}
\examples{
data(Or82a)
}
\references{
\itemize{ \item \link[=reference]{(Hallem and Carlson, 2006)} :
Hallem E. A. & Carlson J. R., 2006, \emph{Coding of odors by a receptor
repertoire}, Cell,125:143-160 \item \link[=reference]{(Hallem et al., 2004)}
: Hallem E.A., Ho M. G. and Carlson J.R., 2004, \emph{The Molecular Basis of
Odor Coding in the Drosophila Antenna}, Cell, 117:965-979 \item
\link[=reference]{(de Bruyne et al., 2001)} : de Bruyne M., Foster K.,
Carlson J. R., 2001, \emph{Odor coding in the Drosophila antenna, Neuron},
30:537-552 \item \link[=reference]{(Kreher et al., 2008)} : Kreher S. A.,
Mathew D., Kim J., Carlson J. R., 2008, \emph{Translation of Sensory Input
into Behavioral Output via an Olfactory System}, Neuron, 59:110-124 }
}
\keyword{datasets}

