% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Or83a}
\alias{Or83a}
\title{Or83a}
\format{A data frame with 251 observations on the following 4 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector with compound ID }
\item{list("CAS")}{a factor with odorant CAS numbers} }}
\description{
Response profile of Or83a
}
\details{
Or83a \describe{ \item{list("Sequence location")}{ 3R:1,223,365..1,224,918
[-] \link[=reference]{(Tweedie et al., 2009)} }\item{:}{
3R:1,223,365..1,224,918 [-] \link[=reference]{(Tweedie et al., 2009)} }
\item{list("Expression")}{ larval specific \link[=reference]{(Couto et al.,
2005)}, \link[=reference]{(Fishilevich et al., 2005)}; \cr Larval specific
odor receptor 83a was detected by in situ hybridization
\link[=reference]{(Fishilevich et al., 2005)}. }\item{:}{ larval specific
\link[=reference]{(Couto et al., 2005)}, \link[=reference]{(Fishilevich et
al., 2005)}; \cr Larval specific odor receptor 83a was detected by in situ
hybridization \link[=reference]{(Fishilevich et al., 2005)}. }
\item{list("Comment")}{ Or83a is located within the odorant binding protein
83 (Obp83) cluster between Obp83b and Obp83c \link[=reference]{(Hekmat-Scafe
et al., 2002)}. }\item{:}{ Or83a is located within the odorant binding
protein 83 (Obp83) cluster between Obp83b and Obp83c
\link[=reference]{(Hekmat-Scafe et al., 2002)}. } }
}
\examples{
data(Or83a)
}
\keyword{datasets}

