% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Or83c}
\alias{Or83c}
\title{Or83c}
\format{A data frame with 251 observations on the following 4 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector with compound ID }
\item{list("CAS")}{a factor with odorant CAS numbers} }}
\description{
Response profile of Or83c
}
\details{
Or83c \describe{ \item{list("Sequence location")}{3R:1,902,526..1,904,005
[-] \link[=reference]{(Tweedie et al., 2009)}
}\item{:}{3R:1,902,526..1,904,005 [-] \link[=reference]{(Tweedie et al.,
2009)} } \item{list("Expression")}{ in adult \link[=reference]{(Couto et
al., 2005)}, \link[=reference]{(Vosshall et al., 2000)} }\item{:}{ in adult
\link[=reference]{(Couto et al., 2005)}, \link[=reference]{(Vosshall et al.,
2000)} } \item{list("Sensilla")}{ trichoid sensilla (antenna)
\link[=reference]{(van der Goes van Naters and Carlson, 2007)}}\item{:}{
trichoid sensilla (antenna) \link[=reference]{(van der Goes van Naters and
Carlson, 2007)}} \item{list("Neuron")}{ at2 \link[=reference]{(Couto et al.,
2005)} }\item{:}{ at2 \link[=reference]{(Couto et al., 2005)} }
\item{list("Coexpression")}{ no recorded }\item{:}{ no recorded }
\item{list("Glomerulus")}{ DC3 \link[=reference]{(Couto et al., 2005)},
\link[=reference]{(Fishilevich and Vosshall, 2005)} }\item{:}{ DC3
\link[=reference]{(Couto et al., 2005)}, \link[=reference]{(Fishilevich and
Vosshall, 2005)} } \item{list("Comment")}{ \enumerate{ \item Or83c failed to
produce responses when tested in the empty neuron, but perhaps the test
neuron fails to have some of the necessary accessory components for the
activity of these specific receptors. For instance, the odorant-binding
protein, Lush \link[=reference]{(Davis, 2007)}.

\item The aT2 sensilla appeared to have response to virgin females extract
\link[=reference]{(van der Goes van Naters and Carlson, 2007)}.  }
}\item{:}{ \enumerate{ \item Or83c failed to produce responses when tested
in the empty neuron, but perhaps the test neuron fails to have some of the
necessary accessory components for the activity of these specific receptors.
For instance, the odorant-binding protein, Lush \link[=reference]{(Davis,
2007)}.

\item The aT2 sensilla appeared to have response to virgin females extract
\link[=reference]{(van der Goes van Naters and Carlson, 2007)}.  } } }
}
\examples{
data(Or83c)
}
\keyword{datasets}

