% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Or85f}
\alias{Or85f}
\title{Or85f}
\format{A data frame with 251 observations on the following 6 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector with compound ID }
\item{list("CAS")}{a factor with odorant CAS numbers}
\item{list("Hallem.2006.EN")}{a numeric vector; electrophysiological
recording in empty neuron from Hallem et al. 2006 \link[=reference]{(Hallem
and Carlson, 2006)} } \item{list("Hallem.2004.EN")}{a numeric vector;
electrophysiological recording in empty neuron from Hallem et al. 2004
\link[=reference]{(Hallem et al., 2004)} } }}
\description{
Response profile of Or85f
}
\details{
Or85f \describe{ \item{list("Sequence location")}{ 3R:5,223,912..5,225,273
[+] \link[=reference]{(Tweedie et al., 2009)} }\item{:}{
3R:5,223,912..5,225,273 [+] \link[=reference]{(Tweedie et al., 2009)} }
\item{list("Expression")}{ in adult \link[=reference]{(Couto et al., 2005)},
\link[=reference]{(Vosshall et al., 2000)} }\item{:}{ in adult
\link[=reference]{(Couto et al., 2005)}, \link[=reference]{(Vosshall et al.,
2000)} } \item{list("Sensilla")}{ small basiconic sensilla (antenna)
\link[=reference]{(Couto et al., 2005)} }\item{:}{ small basiconic sensilla
(antenna) \link[=reference]{(Couto et al., 2005)} } \item{list("Neuron")}{
ab10 \link[=reference]{(Couto et al., 2005)} }\item{:}{ ab10
\link[=reference]{(Couto et al., 2005)} } \item{list("Coexpression")}{ with
Or49a \link[=reference]{(Couto et al., 2005)} }\item{:}{ with Or49a
\link[=reference]{(Couto et al., 2005)} } \item{list("Glomerulus")}{ DL4
\link[=reference]{(Couto et al., 2005)} }\item{:}{ DL4
\link[=reference]{(Couto et al., 2005)} } }
}
\examples{
data(Or85f)
}
\references{
\itemize{ \item \link[=reference]{(Hallem and Carlson, 2006)} :
Hallem E. A. & Carlson J. R., 2006, \emph{Coding of odors by a receptor
repertoire}, Cell,125:143-160 \item \link[=reference]{(Hallem et al., 2004)}
: Hallem E.A., Ho M. G. and Carlson J.R., 2004, \emph{The Molecular Basis of
Odor Coding in the Drosophila Antenna}, Cell, 117:965-979 }
}
\keyword{datasets}

