% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Or92a}
\alias{Or92a}
\title{Or92a}
\format{A data frame with 251 observations on the following 8 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector} \item{list("CAS")}{a factor with
odorant CAS numbers} \item{list("Bruyne.2001.WT")}{a numeric vector;
electrophysiological recording in wild type from de Bruyne et al. 2001
\link[=reference]{(de Bruyne et al., 2001)} }
\item{list("Bruyne.2001.RR")}{a numeric vector; electrophysiological
recording data of response range from de Bruyne et al. 2001
\link[=reference]{(de Bruyne et al., 2001)} }
\item{list("Dobritsa.2003.EN")}{a numeric vector; electrophysiological
recording in empty neuron from Dobritsa et al. 2003
\link[=reference]{(Dobritsa et al., 2003)} }
\item{list("Galizia.2009.nmr")}{a numeric vector; normalized mean values of
calcium imaging recording. } }}
\description{
Response profile of Or92a
}
\details{
Or92a \describe{ \item{list("Sequence location")}{ 3R:16,330,510..16,331,908
[+] \link[=reference]{(Tweedie et al., 2009)} }\item{:}{
3R:16,330,510..16,331,908 [+] \link[=reference]{(Tweedie et al., 2009)} }
\item{list("Expression")}{ in adult \link[=reference]{(Couto et al., 2005)}
}\item{:}{ in adult \link[=reference]{(Couto et al., 2005)} }
\item{list("Sensilla")}{ large basiconic sensilla (antenna)
\link[=reference]{(Couto et al., 2005)} }\item{:}{ large basiconic sensilla
(antenna) \link[=reference]{(Couto et al., 2005)} } \item{list("Neuron")}{
ab1 \link[=reference]{(Jones et al., 2007)}; \cr ab1B;
\link[=reference]{(Root et al., 2007)} \cr }\item{:}{ ab1
\link[=reference]{(Jones et al., 2007)}; \cr ab1B; \link[=reference]{(Root
et al., 2007)} \cr } \item{list("Coexpression")}{ no recorded }\item{:}{ no
recorded } \item{list("Glomerulus")}{ VA2 \link[=reference]{(Couto et al.,
2005)} }\item{:}{ VA2 \link[=reference]{(Couto et al., 2005)} }
\item{list("Comment")}{ \enumerate{ \item Or42b and Or92a expressed in ab1
sensilla neurons have not been conclusively assigned to either ab1A or ab1B
\link[=reference]{(Jones et al., 2007)}.  \item The Or42b receptor is
expressed in ab1A ORNs, that was proved by using Or42b mutation. Hereby it
can be concluded, that the Or92a receptor is expressed in ab1B ORNs.
\link[=reference]{(Root et al., 2007)}.  } }\item{:}{ \enumerate{ \item
Or42b and Or92a expressed in ab1 sensilla neurons have not been conclusively
assigned to either ab1A or ab1B \link[=reference]{(Jones et al., 2007)}.
\item The Or42b receptor is expressed in ab1A ORNs, that was proved by using
Or42b mutation. Hereby it can be concluded, that the Or92a receptor is
expressed in ab1B ORNs. \link[=reference]{(Root et al., 2007)}.  } } }
}
\examples{
data(Or92a)
}
\references{
\itemize{ \item \link[=reference]{(de Bruyne et al., 2001)} : de
Bruyne M., Foster K., Carlson J. R., 2001, \emph{Odor coding in the
Drosophila antenna, Neuron}, 30:537-552 \item \link[=reference]{(Dobritsa et
al., 2003)} : Dobritsa A. A., van der Goes van Naters W., Warr C.G.,
Steinbrecht R. A., Carlson J.R., 2003, \emph{Integrating the molecular and
cellular basis of odor coding in the Drosophila antenna}, Neuron, 37:827-841
\item \link[=reference]{(Root et al., 2007)} : Root C. M., Semmelhack J. L.,
Wong A. M., Flores J. and Wang J. W., 2007, \emph{Propagation of olfactory
information in Drosophila}, PNAS, 104: 11826-11831 }
}
\keyword{datasets}

