\name{ORs}
\Rdversion{1.1}
\alias{ORs}
\docType{data}
\title{
odorant receptor and expression
}
\description{
odorant receptor and expression
}
\usage{data(ORs)}
\format{
  A data frame with 74 observations on the following 2 variables.
  \describe{
    \item{\code{OR}}{a factor with odorant names }
    \item{\code{expression}}{a numeric vector; \code{0}, \code{1}, \code{2} and \code{NA} indicate expression in adult, in larvae, both in adult 
	and larvae and no recorded, respectively.}
  }
}
\examples{
data(ORs)
}
\keyword{datasets}
