% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{reference}
\alias{reference}
\title{reference}
\format{A data frame with 42 observations on the following 5 variables.
\describe{ \item{list("paper")}{a factor with the papers including Author
and published year } \item{list("title")}{ a factor with the title of
studies } \item{list("journal")}{a factor with the journals }
\item{list("volume")}{a factor with the volumes } \item{list("pages")}{a
factor with the pages } }}
\description{
reference for response data
}
\details{
\itemize{ \item (Benton et al., 2006) : Benton R, Sachse S, Michnick SW,
Vosshall LB, (2006), \emph{Atypical membrane topology and heteromeric
function of Drosophila odorant receptors in vivo}, PLoS Biol,4:e20 \item
(Benton et al., 2009) : Benton R, Vannice KS, Gomez-Diaz C, Vosshall LB,
(2009), \emph{Variant ionotropic glutamate receptors as chemosensory
receptors in Drosophila}, Cell,136:149-162 \item (Berdnik et al., 2006) :
Berdnik D, Chihara T, Couto A, Luo L, (2006), \emph{Wiring stability of the
adult Drosophila olfactory circuit after lesion}, J Neurosci 26:3367-3376
\item (Clyne et al., 1999) : Berdnik D, Chihara T, Couto A, Luo L, (1999),
\emph{Wiring stability of the adult Drosophila olfactory circuit after
lesion}, J Neurosci 26:3367-3376 \item (Couto et al., 2005) : Couto A,
Alenius M, Dickson BJ, (2005), \emph{Molecular, anatomical, and functional
organization of the Drosophila olfactory system}, Curr Biol 15:1535--1547
\item (Davis, 2007) : Davis RL, (2007), \emph{The scent of Drosophila sex},
Neuron 54:14-16 \item (de Bruyne et al., 1999) : de Bruyne M, Clyne PJ,
Carlson JR, (1999), \emph{Odor coding in a model olfactory organ: the
Drosophila maxillary palp}, J Neurosci 19:4520-4532 \item (de Bruyne et al.,
2001) : de Bruyne M, Foster K, Carlson JR, (2001), \emph{Odor coding in the
Drosophila antenna}, Neuron 30:537-552 \item (Dobritsa et al., 2003) :
Dobritsa AA, van der Goes van Naters W, Warr CG, Steinbrecht RA, Carlson JR,
(2003), \emph{Integrating the molecular and cellular basis of odor coding in
the Drosophila antenna}, Neuron 37:827-841 \item (Ejima et al., 2007) :
Ejima A, Smith BPC, Lucas C, van der Goes van Naters W, Miller CJ, Carlson
JR, Levine JD, Griffith LC, (2007), \emph{Generalization of courtship
learning in Drosophila is mediated by cis-vaccenyl acetate}, Curr Biol
17:599-605 \item (Fishilevich and Vosshall, 2005) : Fishilevich E, Vosshall
LB, (2005), \emph{Genetic and functional subdivision of the Drosophila
antennal lobe}, Curr Biol 15:1548-1553 \item (Fishilevich et al., 2005) :
Fishilevich E, Domingos AI, Asahina K, Naef F, Vosshall LB, Louis M, (2005),
\emph{Chemotaxis behavior mediated by single larval olfactory neurons in
Drosophila}, Curr Biol 15:2086-2096 \item (Gao et al., 2000) : Gao Q, Yuan
B, Chess A, (2000), \emph{Convergent projections of Drosophila olfactory
neurons to specific glomeruli in the antennal lobe}, Nat Neurosci 3:780-785
\item (Goldman et al., 2005) : Goldman AL, Van der Goes van Naters W,
Lessing D, Warr CG, Carlson JR, (2005), \emph{Coexpression of two functional
odor receptors in one neuron}, Neuron 45:661-666 \item (Ha and Smith, 2006)
: Ha TS, Smith DP, (2006), \emph{A pheromone receptor mediates
11-cis-vaccenyl acetate-induced responses in Drosophila}, J Neurosci
26:8727-8733 \item (Hallem and Carlson, 2006) : Hallem EA, Carlson JR,
(2006), \emph{Coding of odors by a receptor repertoire}, Cell 125:143--160
\item (Hallem et al., 2004) : Hallem EA, Ho MG, Carlson JR, (2004),
\emph{The molecular basis of odor coding in the Drosophila antenna.}, Cell
117:965-979 \item (Hekmat-Scafe et al., 2002) : Hekmat-Scafe DS, Scafe CR,
McKinney AJ, Tanouye MA, (2002), \emph{Genome-wide analysis of the
odorant-binding protein gene family in Drosophila melanogaster}, Genome Res
12:1357-1369 \item Jefferis et al., 2007 : Jefferis GS, Potter CJ, Chan AM,
Marin EC, Rohlfing T, Maurer CR, Jr., Luo L, (2007), \emph{Comprehensive
maps of Drosophila higher olfactory centers: spatially segregated fruit and
pheromone representation}, Cell 128:1187-1203 \item (Jones et al., 2007) :
Jones WD, Cayirlioglu P, Kadow IG, Vosshall LB, (2007), \emph{Two
chemosensory receptors together mediate carbon dioxide detection in
Drosophila}, Nature 445:86--90 \item (Kim et al., 2005) : Kim H, Golub GH,
Park H, (2005), \emph{Missing value estimation for DNA microarray gene
expression data: local least squares imputation}, Bioinformatics 21:187--198
\item (Kondoh et al., 2003) : Kondoh Y, Kaneshiro KY, Kimura K-i, Yamamoto
D, (2003) , \emph{Evolution of sexual dimorphism in the olfactory brain of
Hawaiian Drosophila}, Proc Biol Sci 270:1005-1013 \item (Kreher et al.,
2005) : Kreher SA, Kwon JY, Carlson JR (2005), \emph{The molecular basis of
odor coding in the Drosophila larva}, Neuron 46:445-456 \item (Kreher et
al., 2008) : Kreher SA, Mathew D, Kim J, Carlson JR, (2008),
\emph{Translation of sensory input into behavioral output via an olfactory
system}, Neuron 59:110-124 \item (Kurtovic et al., 2007) : Kurtovic A,
Widmer A, Dickson BJ, (2007), \emph{A single class of olfactory neurons
mediates behavioural responses to a Drosophila sex pheromone}, Nature
446:542-546 \item (Kwon et al., 2007) : Kwon JY, Dahanukar A, Weiss LA,
Carlson JR, (2007), \emph{molecular basis of CO2 reception in Drosophila},
Proc Natl Acad Sci U S A 104:3574--3578 \item (Laissue and Vosshall, 2008) :
Laissue PP, Vosshall LB, (2008), \emph{The Olfactory Sensory Map in
Drosophila. Brain Development in Drosophila melanogaster} \item (Manoli et
al., 2005) : Manoli DS, Foss M, Villella A, Taylor BJ, Hall JC, Baker BS,
(2005), \emph{Male-specific fruitless specifies the neural substrates of
Drosophila courtship behaviour}, Nature 436:395-400 \item (Pelz, 2005) :
Pelz D, (2005), \emph{Functional characterization of Drosophila melanogaster
Olfactory Receptor Neurons}, In: Biologie. Berlin: Freie Universit?t Berlin
\item (Pelz et al., 2006) : Pelz D, Roeske T, Syed Z, de Bruyne M, Galizia
CG, (2006), \emph{The molecular receptive range of an olfactory receptor in
vivo (Drosophila melanogaster Or22a)}, J Neurobiol 66:1544-1563 \item
(Robertson et al., 2003) : Robertson HM, Warr CG, Carlson JR, (2003),
\emph{Molecular evolution of the insect chemoreceptor gene superfamily in
Drosophila melanogaster}, Proc Natl Acad Sci U S A 100 Suppl 2:14537-14542
\item (Root et al., 2007) : Root CM, Semmelhack JL, Wong AM, Flores J, Wang
JW, (2007), \emph{Propagation of olfactory information in Drosophila}, Proc
Natl Acad Sci U S A 104:11826-11831 \item (Schmuker et al., 2007) : Schmuker
M, de Bruyne M, Haehnel M, Schneider G, (2007), \emph{Predicting olfactory
receptor neuron responses from odorant structure}, Chem Cent J 1:11 \item
(Stensmyr et al., 2003) : Stensmyr MC, Giordano E, Balloi A, Angioy AM,
Hansson BS, (2003), \emph{Novel natural ligands for Drosophila olfactory
receptor neurones}, J Exp Biol 206:715-724 \item (Stockinger et al., 2005) :
Stockinger P, Kvitsiani D, Rotkopf S, Tirian L, Dickson BJ, (2005),
\emph{Neural circuitry that governs Drosophila male courtship behavior},
Cell 121:795-807 \item (Suh et al., 2004) : Stockinger P, Kvitsiani D,
Rotkopf S, Tirian L, Dickson BJ, (2005), \emph{Neural circuitry that governs
Drosophila male courtship behavior}, Cell 121:795-807 \item (Turner et al.,
2009) : Turner SL, Ray A, (2009), \emph{Modification of CO2 avoidance
behaviour in Drosophila by inhibitory odorants}, Nature 461:277-281 \item
(Tweedie et al., 2009) : Tweedie S, Ashburner M, Falls K, Leyland P,
McQuilton P, Marygold S, Millburn G, Osumi-Sutherland D, Schroeder A, Seal
R, Zhang H, (2009), \emph{FlyBase: enhancing Drosophila Gene Ontology
annotations}, Nucleic Acids Res 37:D555-559 \item (van der Goes van Naters
and Carlson, 2007) : van der Goes van Naters W, Carlson JR, (2007),
\emph{Receptors and neurons for fly odors in Drosophila}, Curr Biol
17:606-612 \item (Vosshall and Stocker, 2007) : Vosshall LB, Stocker RF,
(2007), \emph{Molecular architecture of smell and taste in Drosophila}, Annu
Rev Neurosci 30:505-533 \item (Vosshall et al., 2000) : Vosshall LB, Wong
AM, Axel R, (2000), \emph{An olfactory sensory map in the fly brain}, Cell
102:147-159 \item (Yao et al., 2005) : Yao CA, Ignell R, Carlson JR, (2005),
\emph{Chemosensory coding by neurons in the coeloconic sensilla of the
Drosophila antenna}, J Neurosci 25:8359--8367 }
}
\examples{
data(reference)
}
\keyword{datasets}

