% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{door_response_matrix}
\alias{door_response_matrix}
\title{DoOR response matrix}
\format{A data frame with 693 observations on 78 variables.}
\description{
globally normalized, precomputed DoOR response matrix
}
\keyword{datasets}

