% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{response.matrix_non.normalized}
\alias{response.matrix_non.normalized}
\title{non normalized DoOR response matrix}
\format{A data frame with 693 observations on 78 variables.}
\description{
precomputed DoOR response matrix with no global normalization applied
}
\keyword{datasets}

