% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{unglobalNorm_response.matrix}
\alias{unglobalNorm_response.matrix}
\title{An unnormalized responses matrix}
\format{A data frame with 251 observations (odors) on the following 67
variables (receptors or ORNs).  \describe{ \item{list("ab2B")}{a numeric
vector} \item{list("ab3B")}{a numeric vector} \item{list("ab4B")}{a numeric
vector} \item{list("ab5B")}{a numeric vector} \item{list("ac1a")}{a numeric
vector} \item{list("ac1b")}{a numeric vector} \item{list("ac2a")}{a numeric
vector} \item{list("ac2b")}{a numeric vector} \item{list("ac3a")}{a numeric
vector} \item{list("ac4")}{a numeric vector} \item{list("Gr21a")}{a numeric
vector} \item{list("Or10a")}{a numeric vector} \item{list("Or13a")}{a
numeric vector} \item{list("Or19a")}{a numeric vector} \item{list("Or1a")}{a
numeric vector} \item{list("Or22a")}{a numeric vector}
\item{list("Or22b")}{a numeric vector} \item{list("Or22c")}{a numeric
vector} \item{list("Or23a")}{a numeric vector} \item{list("Or24a")}{a
numeric vector} \item{list("Or2a")}{a numeric vector} \item{list("Or30a")}{a
numeric vector} \item{list("Or33a")}{a numeric vector}
\item{list("Or33b")}{a numeric vector} \item{list("Or33c")}{a numeric
vector} \item{list("Or35a")}{a numeric vector} \item{list("Or42a")}{a
numeric vector} \item{list("Or42b")}{a numeric vector}
\item{list("Or43a")}{a numeric vector} \item{list("Or43b")}{a numeric
vector} \item{list("Or45a")}{a numeric vector} \item{list("Or45b")}{a
numeric vector} \item{list("Or46a")}{a numeric vector}
\item{list("Or47a")}{a numeric vector} \item{list("Or47b")}{a numeric
vector} \item{list("Or49a")}{a numeric vector} \item{list("Or49b")}{a
numeric vector} \item{list("Or56a")}{a logical vector}
\item{list("Or59a")}{a numeric vector} \item{list("Or59b")}{a numeric
vector} \item{list("Or59c")}{a numeric vector} \item{list("Or63a")}{a
logical vector} \item{list("Or65a")}{a numeric vector}
\item{list("Or67a")}{a numeric vector} \item{list("Or67b")}{a numeric
vector} \item{list("Or67c")}{a numeric vector} \item{list("Or67d")}{a
numeric vector} \item{list("Or69a")}{a logical vector}
\item{list("Or71a")}{a numeric vector} \item{list("Or74a")}{a numeric
vector} \item{list("Or7a")}{a numeric vector} \item{list("Or82a")}{a numeric
vector} \item{list("Or83a")}{a logical vector} \item{list("Or83c")}{a
logical vector} \item{list("Or85a")}{a numeric vector}
\item{list("Or85b")}{a numeric vector} \item{list("Or85c")}{a numeric
vector} \item{list("Or85d")}{a numeric vector} \item{list("Or85e")}{a
numeric vector} \item{list("Or85f")}{a numeric vector}
\item{list("Or88a")}{a numeric vector} \item{list("Or92a")}{a numeric
vector} \item{list("Or94a")}{a numeric vector} \item{list("Or94b")}{a
numeric vector} \item{list("Or98a")}{a numeric vector} \item{list("Or9a")}{a
numeric vector} \item{list("pb2A")}{a numeric vector} }}
\description{
a responses matrix that has not been globally normalized.
}
\details{
The response values has not been normalized, which means that the responses
are in value from 0 to 1.
}
\examples{
data(unglobalNorm_response.matrix)
}
\keyword{datasets}

