\name{weight.globNorm}
\Rdversion{1.1}
\alias{weight.globNorm}
\docType{data}
\title{
weight matrix for global normalizazion
}
\description{
weight matrix for global normalizazion
}
\usage{data(weight.globNorm)}
\format{
  A data frame with 67 observations (receptors) on the following 27 variables (studies).
}
\details{
Value 1 indicates that this receptor has been measured by this study. NA indicates NOT.
}
\keyword{datasets}
