% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{ab2B}
\alias{ab2B}
\title{ab2B}
\format{A data frame with 251 observations on the following 8 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector with compound ID }
\item{list("CAS")}{a factor with odorant CAS numbers}
\item{list("Hallem.2004.WT")}{a numeric vector; electrophysiological
recording in wild type from Hallem et al. 2004 \link[=reference]{(Hallem et
al., 2004)}} \item{list("Stensmyr.2003.WT")}{a numeric vector;
electrophysiological recording in wild type from Sternsmyr et al. 2003
\link[=reference]{(Stensmyr et al., 2003)}}
\item{list("Schmuker.2007.TR")}{a numeric vector; electrophysiological
recording data from Schmuker et al. 2007 \link[=reference]{(Schmuker et al.,
2007)}} \item{list("Dobritsa.2003.WT")}{a numeric vector;
electrophysiological recording in wild type from Dobritsa et al. 2003
\link[=reference]{(Dobritsa et al., 2003)}} }}
\description{
Response profile of ab2B
}
\details{
\describe{ \item{list("coexpression")}{ Or33b coexpress with Or85a in ab2B
neuron and coconvergence to glomerulus DM5 \link[=reference]{(Fishilevich et
al., 2005)}.  }\item{:}{ Or33b coexpress with Or85a in ab2B neuron and
coconvergence to glomerulus DM5 \link[=reference]{(Fishilevich et al.,
2005)}.  } }
}
\examples{
data(ab2B)
}
\references{
\itemize{ \item \link[=reference]{(Stensmyr et al., 2003)} :
Stensmyr M. C., Giordano E., Balloi A., Angioy A. M., Hansson B. S., 2003,
\emph{Novel natural ligands for Drosophila olfactory receptor neurons}, J.
Exp. Biol. 206:715-724 \item \link[=reference]{ (Dobritsa et al., 2003)} :
Dobritsa A. A., van der Goes van Naters W., Warr C.G., Steinbrecht R. A.,
Carlson J.R., 2003, \emph{Integrating the molecular and cellular basis of
odor coding in the Drosophila antenna}, Neuron, 37:827-841 \item
\link[=reference]{(Hallem et al., 2004)} : Hallem E.A., Ho M. G. and Carlson
J.R., 2004, \emph{The Molecular Basis of Odor Coding in the Drosophila
Antenna}, Cell, 117:965-979 \item \link[=reference]{(Schmuker et al., 2007)}
: Schmuker M., de Bruyne M., Haehnel M., Schneider G., 2007,
\emph{Predicting olfactory receptor neuron responses from odorant
structure},Chem Cent J., 1:11 }
}
\keyword{datasets}

