% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{ac4}
\alias{ac4}
\title{ac4}
\format{A data frame with 693 observations on 8 variables.}
\description{
DoOR response data
}
\details{
pooled measurements of the ac4 sensillum
}
\keyword{datasets}

