% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{AL256}
\alias{AL256}
\title{AL256}
\format{A data frame with 422 observations on the following 1495 variables.}
\description{
A 422 by 1495 Figure matrix of Antennal Lobe
}
\examples{
data(AL256)
}
\references{
Vosshall L. B. and Stocker R. F., 2007, \emph{Molecular
Architecture of Smell and Taste in Drosophila}, Annu. Rev. Neurosci.,
30:505-533
}
\keyword{datasets}

