\name{data.format}
\Rdversion{1.1}
\alias{data.format}
\docType{data}
\title{
data format
}
\description{
  the format of response data
}
\usage{data(data.format)}
\format{
  A data frame with 244 observations on the following 4 variables.
  \describe{
    \item{\code{Class}}{a factor with levels \code{} \code{acid} \code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester} \code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid} \code{terpene}}
    \item{\code{Name}}{a factor with odorant names}
    \item{\code{CID}}{a character vector}
    \item{\code{CAS}}{a factor with odorant CAS numbers}
  }
}
\examples{
data(data.format)
}
\keyword{datasets}
