% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{dataset.info}
\alias{dataset.info}
\title{dataset information}
\format{A data frame with 26 observations on the following 13 variables.
\describe{ \item{list("dataset")}{character string; indicating study code}
\item{list("study")}{character string; indicating study name}
\item{list("other.dataset.in.this.study")}{character string; coding another
dataset in this study}
\item{list("other.dataset.in.this.study.2")}{character string; coding
another dataset in this study} \item{list("SFR.reported")}{character string;
indicating whether spontaneous firing rate was reported in this study }
\item{list("SFR.substracted..")}{character string; indicating whether
spontaneous firing rate has been substracted in this study. }
\item{list("technique")}{character string; indicating technique}
\item{list("data.type")}{character string; indicating data type}
\item{list("control")}{character string; indicating control odor}
\item{list("solvents")}{character string; indicating solvent information}
\item{list("concentration")}{character string; indicating odorant
concentration} \item{list("comment")}{character string; containing comments}
\item{list("link")}{character string; indicating url link for this study} }}
\description{
dataset information
}
\examples{
data(dataset.info)
}
\keyword{datasets}

