\name{glo.dist}
\Rdversion{1.1}
\alias{glo.dist}
\docType{data}
\title{
a matrix of distance coefficients between glomeruli
}
\description{
a matrix of distance coefficients between glomeruli
}
\usage{data(glo.dist)}
\format{
  A data frame with 49 observations on the following 49 variables.
  \describe{
    \item{\code{D}}{a logical vector}
    \item{\code{DA1}}{a numeric vector}
    \item{\code{DA2}}{a numeric vector}
    \item{\code{DA3}}{a numeric vector}
    \item{\code{DA4l}}{a numeric vector}
    \item{\code{DA4m}}{a numeric vector}
    \item{\code{DC1}}{a numeric vector}
    \item{\code{DC2}}{a numeric vector}
    \item{\code{DC3}}{a numeric vector}
    \item{\code{DL1}}{a numeric vector}
    \item{\code{DL2}}{a numeric vector}
    \item{\code{DL3}}{a numeric vector}
    \item{\code{DL4}}{a numeric vector}
    \item{\code{DL5}}{a numeric vector}
    \item{\code{DM1}}{a numeric vector}
    \item{\code{DM2}}{a numeric vector}
    \item{\code{DM3}}{a numeric vector}
    \item{\code{DM4}}{a numeric vector}
    \item{\code{DM5}}{a numeric vector}
    \item{\code{DM6}}{a numeric vector}
    \item{\code{DP1l}}{a numeric vector}
    \item{\code{DP1m}}{a numeric vector}
    \item{\code{extra.glomerulus1}}{a numeric vector}
    \item{\code{extra.glomerulus2}}{a numeric vector}
    \item{\code{V}}{a numeric vector}
    \item{\code{VA1d}}{a numeric vector}
    \item{\code{VA1lm}}{a numeric vector}
    \item{\code{VA1v}}{a numeric vector}
    \item{\code{VA2}}{a numeric vector}
    \item{\code{VA3}}{a numeric vector}
    \item{\code{VA4}}{a numeric vector}
    \item{\code{VA5}}{a numeric vector}
    \item{\code{VA6}}{a numeric vector}
    \item{\code{VA7l}}{a numeric vector}
    \item{\code{VA7m}}{a numeric vector}
    \item{\code{VC1}}{a numeric vector}
    \item{\code{VC2}}{a numeric vector}
    \item{\code{VC3}}{a numeric vector}
    \item{\code{VC4}}{a numeric vector}
    \item{\code{VL1}}{a numeric vector}
    \item{\code{VL2a}}{a numeric vector}
    \item{\code{VL2p}}{a numeric vector}
    \item{\code{VM1}}{a numeric vector}
    \item{\code{VM2}}{a numeric vector}
    \item{\code{VM3}}{a numeric vector}
    \item{\code{VM4}}{a numeric vector}
    \item{\code{VM5d}}{a numeric vector}
    \item{\code{VM6}}{a numeric vector}
    \item{\code{VM7}}{a numeric vector}
  }
}
\details{
	The odorant distance coefficients between glomeruli were computed by 
	using the VRML model from http://flybrain.neurobio.arizona.edu
}
\references{
   \itemize{
     \item \ http://flybrain.neurobio.arizona.edu
     \item \Laissue, P. P.; Reiter, C.; Hiesinger, P. R.; Halter, S.; Fischbach, K. F. & Stocker, R. F.,
	1999, \emph{Three-dimensional reconstruction of the antennal lobe in Drosophila melanogaster.}, 
	J Comp Neurol,, 405:543-552
  }
}
\examples{
data(glo.dist)
}
\keyword{datasets}
