% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Gr21a.Gr63a}
\alias{Gr21a.Gr63a}
\title{Gr21a.Gr63a}
\format{A data frame with 693 observations on 10 variables.}
\description{
DoOR response data
}
\details{
data measured from the CO2 responsive sensillum
}
\keyword{datasets}

