% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Ir31a}
\alias{Ir31a}
\title{Ir31a}
\format{A data frame with 251 observations on the following 4 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector} \item{list("CAS")}{a factor with
odorant CAS numbers} }}
\description{
Response profile of Ir31a
}
\details{
Ir31a \describe{ \item{list("Sequence location")}{ 2L:10,406,075..10,408,011
[+] \link[=reference]{(Tweedie et al., 2009)} }\item{:}{
2L:10,406,075..10,408,011 [+] \link[=reference]{(Tweedie et al., 2009)} }
\item{list("Sensilla")}{ antennal coeloconic sensilla (antenna)
\link[=reference]{ (Benton at. al., 2009) } }\item{:}{ antennal coeloconic
sensilla (antenna) \link[=reference]{ (Benton at. al., 2009) } }
\item{list("Neuron")}{ ac1 \link[=reference]{ (Benton at. al., 2009) }
}\item{:}{ ac1 \link[=reference]{ (Benton at. al., 2009) } }
\item{list("Comment")}{ \enumerate{ \item IRs functions as chemosenor that
does not express ORs or Or83b \link[=reference]{ (Benton at. al., 2009) }.
\item Ir31a is sensitive to ammonia \link[=reference]{ (Benton at. al.,
2009) }.  } }\item{:}{ \enumerate{ \item IRs functions as chemosenor that
does not express ORs or Or83b \link[=reference]{ (Benton at. al., 2009) }.
\item Ir31a is sensitive to ammonia \link[=reference]{ (Benton at. al.,
2009) }.  } } }
}
\examples{
data(Ir31a)
}
\references{
\itemize{ \item \link[=reference]{ (Benton at. al., 2009) } : R.
Benton, K. S. Vannice, C. Gomez-Diaz, L. B. Vosshall, \emph{Variant
Ionotropic Glutamate Receptors as Chemosensory Receptors in Drosophila},
2009, cell, 136(1):149-162 }
}
\keyword{datasets}

