% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_door_data.R
\name{load_door_data}
\alias{loadData}
\alias{load_door_data}
\title{load_door_data}
\usage{
load_door_data()
}
\value{
attaches all DoOR data to the main workspace
}
\description{
load all DoOR.data
}
\examples{
<<<<<<< HEAD
# load all data to current workspace
\dontrun{
load_door_data()
}

# it is also possible to load individual data sets using data().
data(Or22a)
=======
\dontrun{
load_door_data()
}
>>>>>>> c4f9f57a7c6f0c023719ea905952e319fb8d425f
}
\author{
Daniel Münch <\email{daniel.muench@uni-konstanz.de}>
}

