% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{odor}
\alias{odor}
\title{Database of Odors}
\format{A data frame with 244 observations on the following 19 variables.
\describe{ \item{list("Class")}{a factor with levels \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names }
\item{list("CID")}{a factor with odorant codes } \item{list("CAS")}{a factor
with odorant CAS numbers} \item{list("Code")}{a numeric vector}
\item{list("Formula")}{a factor with odorant formula} \item{list("MW")}{a
numeric vector; mol weight of odor compounds (g/mol) } \item{list("BP")}{a
numeric vector; boiling point of odor compounds (celsius degree) }
\item{list("MP")}{a numeric vector; melting point of odor compounds (celsius
degree) } \item{list("HLC")}{a numeric vector; Henry's Law Constant of odor
compounds (atm-m3/mol) } \item{list("EG")}{a numeric vector; enthalpy of gas
phase of odor compounds (kJ/mol) } \item{list("EC")}{a numeric vector;
enthalpy of condensed phase of odor compounds (kJ/mol) }
\item{list("HBD")}{a numeric vector; H-bond donor count of odor compounds }
\item{list("HBA")}{a numeric vector; H-bond acceptor count of odor compounds
} \item{list("RotB")}{a numeric vector; rotatable bond count of odor
compounds } \item{list("TPSA")}{a numeric vector; topological polar surface
area of odor compounds } \item{list("Volume")}{a numeric vector; volume of
odor compounds } \item{list("VP.25")}{a numeric vector; vapour pressure of
odor compounds in 25 celsius degree } \item{list("pKa.25")}{a numeric
vector; acid-ionization constant of odor compounds in 25 celsius degree} }}
\source{
The chemical data were obtained from internet, published books and
papers.
}
\description{
Database of Odors
}
\examples{
data(odor)
}
\references{
\itemize{ \item pubchem: \url{http://pubchem.ncbi.nlm.nih.gov/}
\item NIST Chemistry Webbook: \url{http://webbook.nist.gov/chemistry/} \item
Sigma aldrich (Flavors and Fragrances):
\url{http://www.sigmaaldrich.com/catalog/search/TablePage/12916829} \item
Wikipedia: \url{http://en.wikipedia.org/wiki/Main_Page} \item Robert C.
Weast, Melvin J. Astle, William H. Beyer (1986-1987) \emph{handbook of
chemistry and physics}. CRC Press, Inc. Boca Raton, Florida }
}
\keyword{datasets}

