% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{odor}
\alias{odor}
\title{odor}
\format{A data frame with 693 observations on 22 variables.}
\description{
data.frame containing information on the odors in DoOR
}
\keyword{datasets}

