% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{odor}
\alias{odor}
\title{odor}
\format{'data.frame':	693 obs. of  22 variables:
$ Class   : Factor w/ 17 levels "acid","acids",..: NA 13 5 5 5 5 5 5 5 5 ...
$ Name    : Factor w/ 690 levels "11-cis vaccenyl acetate",..: 634 675 240 613 283 239 363 436 458 341 ...
$ InChIKey: Factor w/ 693 levels "ACCRBMDJCPPJDX-UHFFFAOYSA-N",..: 482 612 549 252 548 418 196 577 41 462 ...
$ CID     : Factor w/ 687 levels "1001","10050",..: 686 680 139 15 220 189 483 610 564 468 ...
$ CAS     : Factor w/ 677 levels "1001-45-2","10032-13-0",..: 676 591 212 100 379 586 231 114 62 200 ...
$ InChI   : Factor w/ 693 levels "1S/C10H12O2/c1-2-8-12-10(11)9-6-4-3-5-7-9/h3-7H,2,8H2,1H3",..: 692 687 690 331 382 689 294 526 380 292 ...
$ SMILES  : Factor w/ 693 levels "[C@]123[C@H](C(C)(C)[C@H](C1)C(CC2)=C)CC[C@H]3C",..: 692 674 688 440 439 672 543 340 370 611 ...
$ Code    : Factor w/ 168 levels "","2EBM","2EPM",..: 163 1 1 1 1 1 1 1 1 1 ...
$ Formula : Factor w/ 328 levels "","C10H10O2",..: NA 323 327 155 176 326 130 245 174 129 ...
$ MW      : num  NA 18 35 88.2 102.2 ...
$ BP      : num  NA 100 NA 158 178 -33 170 154 95 7 ...
$ MP      : num  NA 0 NA 25 14 -78 10 -18 -60 -93 ...
$ HLC     : num  NA NA NA 1.82e-09 NA NA 3.25e-08 4.15e-05 NA 1.77e-05 ...
$ EG      : num  NA -242 NA NA NA ...
$ EC      : num  NA -286 NA NA NA ...
$ HBD     : int  NA 1 2 2 2 1 2 1 1 1 ...
$ HBA     : int  NA 1 1 2 2 1 2 1 1 1 ...
$ RotB    : int  NA 0 0 3 4 0 1 5 2 0 ...
$ TPSA    : num  NA 1 2 52 52 1 46.3 26 26 12 ...
$ Volume  : num  NA 20.1 NA 98.6 116.5 ...
$ VP.25   : num  NA 23.54 NA NA 4.12 ...
$ pKa.25  : num  NA NA NA NA 10.8 ...}
\description{
A data.frame containig information like chemical identifiers and some chemical descriptors for
all odorants in DoOR.
}
\keyword{DoOR}
\keyword{dataset}
\keyword{odor}
\keyword{responding_unit}

