\name{OGN}
\Rdversion{1.1}
\alias{OGN}
\docType{data}
\title{
Mapping of Receptors 
}
\description{
  The map of Receptor to sensillium, receptor neuron and glomerulus
}
\usage{data(OGN)}
\format{
  A data frame with 84 observations on the following 12 variables.
  \describe{
    \item{\code{Receptor}}{a factor with receptor or neuron names}
    \item{\code{Coexpress}}{a factor with coespression}
    \item{\code{Coexpressing1}}{a factor with coespression1}
    \item{\code{Coexpressing2}}{a factor with coespression2}
    \item{\code{Coexpressing3}}{a factor with coespression3}
    \item{\code{Ors}}{a factor with receptor names}
    \item{\code{Sensillium}}{a factor with sensillum type}
    \item{\code{ORN}}{a factor with neuron names}
    \item{\code{Glomerulus}}{a factor with glomerulus names}
    \item{\code{Code}}{a numeric vector}
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
  The data frame is used for showing functional antennal lobe. 
}
\examples{
data(OGN)
}
\keyword{datasets}
