% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Or33b}
\alias{Or33b}
\title{Or33b}
\format{A data frame with 251 observations on the following 7 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector with compound ID }
\item{list("CAS")}{a factor with odorant CAS numbers}
\item{list("Hallem.2006.EN")}{a numeric vector; electrophysiological
recording in empty neuron from Hallem et al. 2006 \link[=reference]{(Hallem
and Carlson, 2006)} } \item{list("Hallem.2004.EN")}{a numeric vector;
electrophysiological recording in empty neuron from Hallem et al. 2004
\link[=reference]{(Hallem et al., 2004)} } \item{list("Kreher.2008.EN")}{a
numeric vector; electrophysiological recording in empty neuron from Kreher
et al. 2008 \link[=reference]{(Kreher et al., 2008)} } }}
\description{
Response profile of Or33b
}
\details{
Or33b \describe{ \item{list("Sequence location")}{ 2L:11,936,181..11,937,501
[+] \link[=reference]{(Tweedie et al., 2009)} }\item{:}{
2L:11,936,181..11,937,501 [+] \link[=reference]{(Tweedie et al., 2009)} }
\item{list("Expression")}{ in adult \link[=reference]{(Vosshall et al.,
2000)}, \link[=reference]{(Fishilevich and Vosshall, 2005)},
\link[=reference]{(Stockinger et al., 2005)} and larvae
\link[=reference]{(Couto et al., 2005)}, \link[=reference]{(Fishilevich et
al., 2005)} }\item{:}{ in adult \link[=reference]{(Vosshall et al., 2000)},
\link[=reference]{(Fishilevich and Vosshall, 2005)},
\link[=reference]{(Stockinger et al., 2005)} and larvae
\link[=reference]{(Couto et al., 2005)}, \link[=reference]{(Fishilevich et
al., 2005)} } \item{list("Sensilla")}{ thin basiconic sensilla (antenna)
\link[=reference]{(Couto et al., 2005)} }\item{:}{ thin basiconic sensilla
(antenna) \link[=reference]{(Couto et al., 2005)} } \item{list("Neuron")}{
\cr Or33b coexpress with Or47a in ab5B neuron \link[=reference]{(Kwon et
al., 2007)}.\cr Or33b coexpress with Or85a in ab2B neuron
\link[=reference]{(Fishilevich and Vosshall, 2005)}. }\item{:}{ \cr Or33b
coexpress with Or47a in ab5B neuron \link[=reference]{(Kwon et al.,
2007)}.\cr Or33b coexpress with Or85a in ab2B neuron
\link[=reference]{(Fishilevich and Vosshall, 2005)}. }
\item{list("Coexpression")}{ with Or47a \link[=reference]{(Kwon et al.,
2007)} and with Or85a \link[=reference]{(Fishilevich and Vosshall, 2005)}
\cr Or33b and Or47a coexpress along with Or83b in larvae
\link[=reference]{(Fishilevich et al., 2005)}. Or33b is coexpressed with
both Or47a and Or85a. Or47a and Or85a are expressed in different ORNs
\link[=reference]{(Fishilevich and Vosshall, 2005)}. }\item{:}{ with Or47a
\link[=reference]{(Kwon et al., 2007)} and with Or85a
\link[=reference]{(Fishilevich and Vosshall, 2005)} \cr Or33b and Or47a
coexpress along with Or83b in larvae \link[=reference]{(Fishilevich et al.,
2005)}. Or33b is coexpressed with both Or47a and Or85a. Or47a and Or85a are
expressed in different ORNs \link[=reference]{(Fishilevich and Vosshall,
2005)}. } \item{list("Glomerulus")}{ DM3 \link[=reference]{(Fishilevich and
Vosshall, 2005)} + DM5 \link[=reference]{(Fishilevich and Vosshall, 2005)}
\cr Or33b and Or47a coconvergence to glomerulus DM3
\link[=reference]{(Fishilevich and Vosshall, 2005)}. \cr Or33b and Or85a
coconvergence to glomerulus DM5 \link[=reference]{(Fishilevich and Vosshall,
2005)}. }\item{:}{ DM3 \link[=reference]{(Fishilevich and Vosshall, 2005)} +
DM5 \link[=reference]{(Fishilevich and Vosshall, 2005)} \cr Or33b and Or47a
coconvergence to glomerulus DM3 \link[=reference]{(Fishilevich and Vosshall,
2005)}. \cr Or33b and Or85a coconvergence to glomerulus DM5
\link[=reference]{(Fishilevich and Vosshall, 2005)}. } }
}
\examples{
data(Or33b)
}
\references{
\itemize{ \item \link[=reference]{(Hallem and Carlson, 2006)} :
Hallem E. A. & Carlson J. R., 2006, \emph{Coding of odors by a receptor
repertoire}, Cell,125:143-160 \item \link[=reference]{(Hallem et al., 2004)}
: Hallem E.A., Ho M. G. and Carlson J.R., 2004, \emph{The Molecular Basis of
Odor Coding in the Drosophila Antenna}, Cell, 117:965-979 \item
\link[=reference]{(Kreher et al., 2008)} : Kreher S. A., Mathew D., Kim J.,
Carlson J. R., 2008, \emph{Translation of Sensory Input into Behavioral
Output via an Olfactory System}, Neuron, 59:110-124 }
}
\keyword{datasets}

