% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Or33c}
\alias{Or33c}
\title{Or33c}
\format{A data frame with 251 observations on the following 5 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector with compound ID }
\item{list("CAS")}{a factor with odorant CAS numbers}
\item{list("Goldman.2005.EN")}{a numeric vector; electrophysiological
recording in empty neuron from Goldman et al. 2005
\link[=reference]{(Goldman et al., 2005)} } }}
\description{
Response profile of Or33c
}
\details{
Or33c \describe{ \item{list("Sequence location")}{ 2L:11,937,969..11,939,202
[+] \link[=reference]{(Tweedie et al., 2009)} }\item{:}{
2L:11,937,969..11,939,202 [+] \link[=reference]{(Tweedie et al., 2009)} }
\item{list("Expression")}{ in adult \link[=reference]{(Couto et al., 2005)},
\link[=reference]{(Vosshall et al., 2000)}, \link[=reference]{(Fishilevich
and Vosshall, 2005)} }\item{:}{ in adult \link[=reference]{(Couto et al.,
2005)}, \link[=reference]{(Vosshall et al., 2000)},
\link[=reference]{(Fishilevich and Vosshall, 2005)} }
\item{list("Sensilla")}{ thin basiconic sensilla (palp)
\link[=reference]{(Goldman et al., 2005)}, \link[=reference]{(Couto et al.,
2005)} }\item{:}{ thin basiconic sensilla (palp) \link[=reference]{(Goldman
et al., 2005)}, \link[=reference]{(Couto et al., 2005)} }
\item{list("Neuron")}{ pb2A \link[=reference]{(Goldman et al., 2005)}
}\item{:}{ pb2A \link[=reference]{(Goldman et al., 2005)} }
\item{list("Coexpression")}{ with Or85e \link[=reference]{(Couto et al.,
2005)} }\item{:}{ with Or85e \link[=reference]{(Couto et al., 2005)} }
\item{list("Glomerulus")}{ VC1 \link[=reference]{(Couto et al., 2005)}
}\item{:}{ VC1 \link[=reference]{(Couto et al., 2005)} } }
}
\examples{
data(Or33c)
}
\references{
\itemize{ \item \link[=reference]{(Goldman et al., 2005)} :
Goldman A. L., Van der Goes van Naters W., Lessing D, Warr C. G., Carlson J.
R., 2005, \emph{Coexpression of Two Functional Odor Receptors in one
Neuron}, Neuron, 45:661-666 }
}
\keyword{datasets}

