% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Or63a}
\alias{Or63a}
\title{Or63a}
\format{A data frame with 251 observations on the following 4 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector with compound ID }
\item{list("CAS")}{a factor with odorant CAS numbers} }}
\description{
Response profile of Or63a
}
\details{
Or63a \describe{ \item{list("Sequence location")}{ 3L:2,961,744..2,963,616
[+] \link[=reference]{(Tweedie et al., 2009)} }\item{:}{
3L:2,961,744..2,963,616 [+] \link[=reference]{(Tweedie et al., 2009)} }
\item{list("Expression")}{ larval specific \link[=reference]{(Kreher et al.,
2005)}, \link[=reference]{(Fishilevich et al., 2005)}; \cr Larval specific
odor receptor 63a was detected by in situ hybridization
\link[=reference]{(Fishilevich et al., 2005)}.  Or63a was not detected in
adult olfactory organs by in situ hybridization analysis
\link[=reference]{(Vosshall et al., 2000)}.}\item{:}{ larval specific
\link[=reference]{(Kreher et al., 2005)}, \link[=reference]{(Fishilevich et
al., 2005)}; \cr Larval specific odor receptor 63a was detected by in situ
hybridization \link[=reference]{(Fishilevich et al., 2005)}.  Or63a was not
detected in adult olfactory organs by in situ hybridization analysis
\link[=reference]{(Vosshall et al., 2000)}.} }
}
\examples{
data(Or63a)
}
\keyword{datasets}

