\name{Or69a}
\Rdversion{1.1}
\alias{Or69a}
\docType{data}
\title{
Or69a
}
\description{
  Response profile of Or69a
}
\usage{data(Or69a)}
\format{
  A data frame with 251 observations on the following 4 variables.
  \describe{
    \item{\code{Class}}{a factor with levels \code{} \code{acid} \code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester} \code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid} \code{terpene}}
    \item{\code{Name}}{a factor with odorant names}
    \item{\code{CID}}{a character vector with compound ID }
    \item{\code{CAS}}{a factor with odorant CAS numbers}
  }
}
\details{
  Or69a
  \describe{
     \item{\code{Sequence location}:}{3L:12,957,460..12,960,833 [-] \link[=reference]{(Tweedie et al., 2009)} } 
     \item{\code{Expression}:}{ in adult \link[=reference]{(Couto et al., 2005)} }   
     \item{\code{Sensilla}:}{ small basiconic sensilla (antenna) (antenna) \link[=reference]{(Couto et al., 2005)}} 
     \item{\code{Neuron}:}{ ab9 \link[=reference]{(Couto et al., 2005)} } 
     \item{\code{Coexpression}:}{ Or69aA and Or69aB coexpress in ab9, coconvergent to glomerulus D \link[=reference]{(Couto et al., 2005)}. } 
     \item{\code{Glomerulus}:}{ D \link[=reference]{(Couto et al., 2005)} }
     \item{\code{Isoform}:}{ Or69aA and Or69aB \link[=reference]{(Couto et al., 2005)} }
  }
}
\examples{
data(Or69a)
}
\keyword{datasets}
