\name{Or83a}
\Rdversion{1.1}
\alias{Or83a}
\docType{data}
\title{
Or83a
}
\description{
  Response profile of Or83a
}
\usage{data(Or83a)}
\format{
  A data frame with 251 observations on the following 4 variables.
  \describe{
    \item{\code{Class}}{a factor with levels \code{} \code{acid} \code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester} \code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid} \code{terpene}}
    \item{\code{Name}}{a factor with odorant names}
    \item{\code{CID}}{a character vector with compound ID }
    \item{\code{CAS}}{a factor with odorant CAS numbers}
  }
}
\details{
  Or83a
  \describe{
     \item{\code{Sequence location}:}{ 3R:1,223,365..1,224,918 [-] \link[=reference]{(Tweedie et al., 2009)} } 
     \item{\code{Expression}:}{ larval specific \link[=reference]{(Couto et al., 2005)}, \link[=reference]{(Fishilevich et al., 2005)}; \cr Larval 
	specific odor receptor 83a was detected by in situ hybridization \link[=reference]{(Fishilevich et al., 2005)}. }
     \item{\code{Comment}:}{ Or83a is located within the odorant binding protein 83 (Obp83) cluster between Obp83b and Obp83c \link[=reference]{(Hekmat-Scafe et al., 2002)}. }
  }
}
\examples{
data(Or83a)
}
\keyword{datasets}
