% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Or88a}
\alias{Or88a}
\title{Or88a}
\format{A data frame with 251 observations on the following 6 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector} \item{list("CAS")}{a factor with
odorant CAS numbers} \item{list("Hallem.2006.EN")}{a numeric vector;
electrophysiological recording in empty neuron from Hallem et al. 2006
\link[=reference]{(Hallem and Carlson, 2006)} }
\item{list("Hallem.2004.EN")}{a numeric vector; electrophysiological
recording in empty neuron from Hallem et al. 2004 \link[=reference]{(Hallem
et al., 2004)} } }}
\description{
Response profile of Or88a
}
\details{
Or88a \describe{ \item{list("Sequence location")}{ 3R:9,935,167..9,936,545
[+] \link[=reference]{(Tweedie et al., 2009)} }\item{:}{
3R:9,935,167..9,936,545 [+] \link[=reference]{(Tweedie et al., 2009)} }
\item{list("Expression")}{ in adult \link[=reference]{(Couto et al., 2005)},
\link[=reference]{[23]} and larvae \link[=reference]{(Kreher et al., 2005)}
}\item{:}{ in adult \link[=reference]{(Couto et al., 2005)},
\link[=reference]{[23]} and larvae \link[=reference]{(Kreher et al., 2005)}
} \item{list("Sensilla")}{ trichoid sensilla (antenna)
\link[=reference]{(van der Goes van Naters and Carlson, 2007)},
\link[=reference]{(Couto et al., 2005)} }\item{:}{ trichoid sensilla
(antenna) \link[=reference]{(van der Goes van Naters and Carlson, 2007)},
\link[=reference]{(Couto et al., 2005)} } \item{list("Neuron")}{ at4
\link[=reference]{(Couto et al., 2005)} }\item{:}{ at4
\link[=reference]{(Couto et al., 2005)} } \item{list("Coexpression")}{ no
recorded }\item{:}{ no recorded } \item{list("Glomerulus")}{ VA1d
\link[=reference]{(Couto et al., 2005)}, \link[=reference]{[30]} }\item{:}{
VA1d \link[=reference]{(Couto et al., 2005)}, \link[=reference]{[30]} }
\item{list("Comment")}{ Or88a was showed weak response to male cuticular
extracts (action potential 50 <= n < 100 impulses/s), medial response to
virgin female extracts, male genital material and mated female genital
material (action potential 100 <= n < 150 impulses/s) but not virgin-female
genital material and 11-cis-vaccenyl acetate (action potential n < 50
impulses/s) \link[=reference]{(van der Goes van Naters and Carlson, 2007)}.
}\item{:}{ Or88a was showed weak response to male cuticular extracts (action
potential 50 <= n < 100 impulses/s), medial response to virgin female
extracts, male genital material and mated female genital material (action
potential 100 <= n < 150 impulses/s) but not virgin-female genital material
and 11-cis-vaccenyl acetate (action potential n < 50 impulses/s)
\link[=reference]{(van der Goes van Naters and Carlson, 2007)}. } }
}
\examples{
data(Or88a)
}
\references{
\itemize{ \item \link[=reference]{(Hallem and Carlson, 2006)} :
Hallem E. A. & Carlson J. R., 2006, \emph{Coding of odors by a receptor
repertoire}, Cell,125:143-160 \item \link[=reference]{(Hallem et al., 2004)}
: Hallem E.A., Ho M. G. and Carlson J.R., 2004, \emph{The Molecular Basis of
Odor Coding in the Drosophila Antenna}, Cell, 117:965-979 }
}
\keyword{datasets}

