% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Or88a}
\alias{Or88a}
\title{Or88a}
\format{'data.frame':	693 obs. of  8 variables:
$ Class         : Factor w/ 17 levels "acid","acids",..: NA 13 5 5 5 5 5 5 5 5 ...
$ Name          : Factor w/ 690 levels "11-cis vaccenyl acetate",..: 634 675 240 613 283 239 363 436 458 341 ...
$ InChIKey      : Factor w/ 693 levels "ACCRBMDJCPPJDX-UHFFFAOYSA-N",..: 482 612 549 252 548 418 196 577 41 462 ...
$ CID           : Factor w/ 687 levels "1001","10050",..: 686 680 139 15 220 189 483 610 564 468 ...
$ CAS           : Factor w/ 677 levels "1001-45-2","10032-13-0",..: 676 591 212 100 379 586 231 114 62 200 ...
$ Hallem.2006.EN: int  26 NA 26 20 26 NA NA NA NA NA ...
$ Dweck.2015b.WT: num  0 NA NA NA NA NA NA NA NA NA ...
$ Hallem.2004.EN: num  0 NA NA NA NA NA NA NA NA NA ...}
\description{
DoOR response data for responding unit Or88a. Please find detailed
information on the respective sources of the data in door_dataset_info.
}
\keyword{DoOR}
\keyword{Or88a}
\keyword{dataset}
\keyword{responding_unit}
