% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Or98a}
\alias{Or98a}
\title{Or98a}
\format{A data frame with 251 observations on the following 9 variables.
\describe{ \item{list("Class")}{a factor with levels \code{} \code{acid}
\code{alcohol} \code{aldehyde} \code{amine} \code{arom} \code{ester}
\code{ketone} \code{N ring} \code{O ring} \code{other} \code{sulfid}
\code{terpene}} \item{list("Name")}{a factor with odorant names}
\item{list("CID")}{a character vector} \item{list("CAS")}{a factor with
odorant CAS numbers} \item{list("Hallem.2006.EN")}{a numeric vector;
electrophysiological recording in empty neuron from Hallem et al. 2006
\link[=reference]{(Hallem and Carlson, 2006)} }
\item{list("Hallem.2004.EN")}{a numeric vector; electrophysiological
recording in empty neuron from Hallem et al. 2004 \link[=reference]{(Hallem
et al., 2004)} } \item{list("Hallem.2004.WT")}{a numeric vector;
electrophysiological recording in wild type from Hallem et al. 2004
\link[=reference]{(Hallem et al., 2004)} } \item{list("Bruyne.2001.WT")}{a
numeric vector; electrophysiological recording in wild type from de Bruyne
et al. 2001 \link[=reference]{(de Bruyne et al., 2001)} }
\item{list("Bruyne.2001.RR")}{a numeric vector; electrophysiological
recording data of response range from de Bruyne et al. 2001
\link[=reference]{(de Bruyne et al., 2001)} } }}
\description{
Response profile of Or98a
}
\details{
Or98a \describe{ \item{list("Sequence location")}{ 3R:23,671,672..23,672,982
[+] \link[=reference]{(Tweedie et al., 2009)} }\item{:}{
3R:23,671,672..23,672,982 [+] \link[=reference]{(Tweedie et al., 2009)} }
\item{list("Expression")}{ in adult \link[=reference]{(Couto et al., 2005)},
\link[=reference]{(Vosshall et al., 2000)} }\item{:}{ in adult
\link[=reference]{(Couto et al., 2005)}, \link[=reference]{(Vosshall et al.,
2000)} } \item{list("Sensilla")}{ small basiconic sensilla (antenna)
\link[=reference]{(Couto et al., 2005)} }\item{:}{ small basiconic sensilla
(antenna) \link[=reference]{(Couto et al., 2005)} } \item{list("Neuron")}{
ab7A \link[=reference]{(Hallem et al., 2004)} }\item{:}{ ab7A
\link[=reference]{(Hallem et al., 2004)} } \item{list("Coexpression")}{ no
recorded }\item{:}{ no recorded } \item{list("Glomerulus")}{ VM5v
\link[=reference]{(Couto et al., 2005)}, VM5 \link[=reference]{(Hallem et
al., 2004)} }\item{:}{ VM5v \link[=reference]{(Couto et al., 2005)}, VM5
\link[=reference]{(Hallem et al., 2004)} } }
}
\examples{
data(Or98a)
}
\references{
\itemize{ \item \link[=reference]{(Hallem and Carlson, 2006)} :
Hallem E. A. & Carlson J. R., 2006, \emph{Coding of odors by a receptor
repertoire}, Cell,125:143-160 \item \link[=reference]{(Hallem et al., 2004)}
: Hallem E.A., Ho M. G. and Carlson J.R., 2004, \emph{The Molecular Basis of
Odor Coding in the Drosophila Antenna}, Cell, 117:965-979 \item
\link[=reference]{(de Bruyne et al., 2001)} : de Bruyne M., Foster K.,
Carlson J. R., 2001, \emph{Odor coding in the Drosophila antenna, Neuron},
30:537-552 }
}
\keyword{datasets}

