% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{Or98a}
\alias{Or98a}
\title{Or98a}
\format{'data.frame':	693 obs. of  11 variables:
$ Class           : Factor w/ 17 levels "acid","acids",..: NA 13 5 5 5 5 5 5 5 5 ...
$ Name            : Factor w/ 690 levels "11-cis vaccenyl acetate",..: 634 675 240 613 283 239 363 436 458 341 ...
$ InChIKey        : Factor w/ 693 levels "ACCRBMDJCPPJDX-UHFFFAOYSA-N",..: 482 612 549 252 548 418 196 577 41 462 ...
$ CID             : Factor w/ 687 levels "1001","10050",..: 686 680 139 15 220 189 483 610 564 468 ...
$ CAS             : Factor w/ 677 levels "1001-45-2","10032-13-0",..: 676 591 212 100 379 586 231 114 62 200 ...
$ Hallem.2006.EN  : int  12 NA 36 29 33 NA NA NA NA NA ...
$ Bruyne.2001.RR  : int  NA NA NA 0 NA 0 0 NA NA NA ...
$ Bruyne.2001.WT  : num  11 NA NA NA NA NA NA NA NA NA ...
$ Marshall.2010.WT: num  0 14.33 5.17 6.4 NA ...
$ Hallem.2004.EN  : num  NA NA NA NA NA NA NA NA NA NA ...
$ Hallem.2004.WT  : num  NA NA NA NA NA NA NA NA NA NA ...}
\description{
DoOR response data for responding unit Or98a. Please find detailed
information on the respective sources of the data in door_dataset_info.
}
\keyword{DoOR}
\keyword{Or98a}
\keyword{dataset}
\keyword{responding_unit}
