% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{ORs}
\alias{ORs}
\title{odorant receptor and expression}
\format{A data frame with 74 observations on the following 2 variables.
\describe{ \item{list("OR")}{a factor with odorant names }
\item{list("expression")}{a numeric vector; \code{0}, \code{1}, \code{2} and
\code{NA} indicate expression in adult, in larvae, both in adult and larvae
and no recorded, respectively.} }}
\description{
odorant receptor and expression
}
\examples{
data(ORs)
}
\keyword{datasets}

