% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{ORs}
\alias{ORs}
\title{ORs}
\format{
'data.frame':	78 obs. of  1 variables:
$ OR        : Factor w/ 78 levels "ab2B","ab4B",..: 5 6 9 10 12 13 15 29 35 63 ...
}
\description{
A data frame containing the names of all the responding units of that data is existing in DoOR.
}
\keyword{DoOR}
\keyword{ORs}
\keyword{dataset}
\keyword{responding_unit}
