% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{ORs}
\alias{ORs}
\title{ORs}
\format{A data frame with 78 observations on 2 variables.}
\description{
list of responding units in DoOR
}
\details{
including information on whether a unitresponding unit is expressed in adult (0), lava (1) or both (2)
}
\keyword{datasets}

