% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{response.matrix}
\alias{response.matrix}
\title{DoOR response matrix}
\format{A data frame with 693 observations on 78 variables.}
\description{
globally normalized, precomputed DoOR response matrix
}
\keyword{datasets}

