% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{response.matrix}
\alias{response.matrix}
\title{An global normalized responses matrix}
\format{A data frame with 251 observations (odors) on the following 67
variables (receptors or ORNs).  \describe{ \item{list("ab2B")}{a numeric
vector} \item{list("ab3B")}{a numeric vector} \item{list("ab4B")}{a numeric
vector} \item{list("ab5B")}{a numeric vector} \item{list("ac1a")}{a numeric
vector} \item{list("ac1b")}{a numeric vector} \item{list("ac2a")}{a numeric
vector} \item{list("ac2b")}{a numeric vector} \item{list("ac3a")}{a numeric
vector} \item{list("ac4")}{a numeric vector} \item{list("Gr21a")}{a numeric
vector} \item{list("Or10a")}{a numeric vector} \item{list("Or13a")}{a
numeric vector} \item{list("Or19a")}{a numeric vector} \item{list("Or1a")}{a
numeric vector} \item{list("Or22a")}{a numeric vector}
\item{list("Or22b")}{a numeric vector} \item{list("Or22c")}{a numeric
vector} \item{list("Or23a")}{a numeric vector} \item{list("Or24a")}{a
numeric vector} \item{list("Or2a")}{a numeric vector} \item{list("Or30a")}{a
numeric vector} \item{list("Or33a")}{a numeric vector}
\item{list("Or33b")}{a numeric vector} \item{list("Or33c")}{a numeric
vector} \item{list("Or35a")}{a numeric vector} \item{list("Or42a")}{a
numeric vector} \item{list("Or42b")}{a numeric vector}
\item{list("Or43a")}{a numeric vector} \item{list("Or43b")}{a numeric
vector} \item{list("Or45a")}{a numeric vector} \item{list("Or45b")}{a
numeric vector} \item{list("Or46a")}{a numeric vector}
\item{list("Or47a")}{a numeric vector} \item{list("Or47b")}{a numeric
vector} \item{list("Or49a")}{a numeric vector} \item{list("Or49b")}{a
numeric vector} \item{list("Or56a")}{a logical vector}
\item{list("Or59a")}{a numeric vector} \item{list("Or59b")}{a numeric
vector} \item{list("Or59c")}{a numeric vector} \item{list("Or63a")}{a
logical vector} \item{list("Or65a")}{a numeric vector}
\item{list("Or67a")}{a numeric vector} \item{list("Or67b")}{a numeric
vector} \item{list("Or67c")}{a numeric vector} \item{list("Or67d")}{a
numeric vector} \item{list("Or69a")}{a logical vector}
\item{list("Or71a")}{a numeric vector} \item{list("Or74a")}{a numeric
vector} \item{list("Or7a")}{a numeric vector} \item{list("Or82a")}{a numeric
vector} \item{list("Or83a")}{a logical vector} \item{list("Or83c")}{a
logical vector} \item{list("Or85a")}{a numeric vector}
\item{list("Or85b")}{a numeric vector} \item{list("Or85c")}{a numeric
vector} \item{list("Or85d")}{a numeric vector} \item{list("Or85e")}{a
numeric vector} \item{list("Or85f")}{a numeric vector}
\item{list("Or88a")}{a numeric vector} \item{list("Or92a")}{a numeric
vector} \item{list("Or94a")}{a numeric vector} \item{list("Or94b")}{a
numeric vector} \item{list("Or98a")}{a numeric vector} \item{list("Or9a")}{a
numeric vector} \item{list("pb2A")}{a numeric vector} }}
\description{
An global normalized responses matrix weighted by receptor max to study max
ratio
}
\details{
The response values has been normalized. Because some data contribute more
than others based on how many receptors and odors were measured, so that we
introduce a weighted mean to scale the model response values. If all weights
are equal, then the weighted mean is the same as the arithmetic mean.  The
weighted receptor max to study max ratio: weight(RMAX/SMAX) = (
sum(n.receptor[i]*(RMAX[i]/SMAX[i])) + sum(n.odor[i]*(RMAX[i]/SMAX[i])) ) /
( sum(n.receptor [i]) + sum(n.odor [i]) )

consensus.value(global) = consensus.value(unglobalNorm) * weight(RMAX/SMAX)

where RMAX i is the maximum response for target receptor in study i; SMAX i
is the maximum response for all receptor in study i. n.receptor i and n.odor
i are the numbers of receptors and odors that have been measured in study i,
respectively.
}
\examples{
data(response.matrix)
}
\keyword{datasets}

