\name{response.range}
\Rdversion{1.1}
\alias{response.range}
\docType{data}
\title{
response range
}
\description{
response ranges for each study
}
\usage{data(response.range)}
\format{
  A data frame with 27 observations on the following 4 variables.
  \describe{
    \item{\code{study}}{a factor with study names }
    \item{\code{min}}{a numeric vector; minimal response }
    \item{\code{max}}{a numeric vector; maximal response }
    \item{\code{n_odors}}{a numeric vector; number of odors }
  }
}
\examples{
data(response.range)
}
\keyword{datasets}
