% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{response.range}
\alias{response.range}
\title{response.range}
\format{A data frame with 42 observations on 4 variables.}
\description{
the response range of the studies included in DoOR
}
\keyword{datasets}

