% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{response.range}
\alias{response.range}
\title{response range}
\format{A data frame with 27 observations on the following 4 variables.
\describe{ \item{list("study")}{a factor with study names }
\item{list("min")}{a numeric vector; minimal response } \item{list("max")}{a
numeric vector; maximal response } \item{list("n_odors")}{a numeric vector;
number of odors } }}
\description{
response ranges for each study
}
\examples{
data(response.range)
}
\keyword{datasets}

