% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{weight.globNorm}
\alias{weight.globNorm}
\title{weight matrix for global normalizazion}
\format{A data frame with 67 observations (receptors) on the following 27
variables (studies).}
\description{
weight matrix for global normalizazion
}
\details{
Value 1 indicates that this receptor has been measured by this study. NA
indicates NOT.
}
\keyword{datasets}

