% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoOR.data-package.R
\docType{data}
\name{weight.globNorm}
\alias{weight.globNorm}
\title{weight.globNorm}
\format{A data frame with 78 observations on 42 variables.}
\description{
matrix of studies vs responding units, used for calculating the weight for global normalization
}
\keyword{datasets}

